% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\name{AddControlInfo}
\alias{AddControlInfo}
\title{AddControlInfo
Called by ascend object creation. Adds the control information to the dataframe.
This is a hidden function and shouldn't be used directly by users.}
\usage{
AddControlInfo(gene.information, controls)
}
\arguments{
\item{gene.information}{A dataframe containing gene identiers in the first 
column}

\item{controls}{List of controls supplied by the user. They must match the
gene identifiers used in the expression matrix.}
}
\description{
AddControlInfo
Called by ascend object creation. Adds the control information to the dataframe.
This is a hidden function and shouldn't be used directly by users.
}
