% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\docType{methods}
\name{addGeneLabel}
\alias{addGeneLabel}
\title{addGeneLabel}
\usage{
addGeneLabel(x, ..., gene)

\S4method{addGeneLabel}{EMSet}(x, ..., gene = c())
}
\description{
Labels cells if they express gene(s) as a condition.
}
\examples{
\dontrun{
em_set <- addGeneLabel(em_set, gene = c("MALAT1"))
}
}
