% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{applyRLE}
\alias{applyRLE}
\title{Apply size factors to dense matrix}
\usage{
applyRLE(x, y)
}
\arguments{
\item{x}{A dense matrix}

\item{y}{Size factors}
}
\description{
Apply size factors to dense matrix
}
