% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_cellranger.R
\name{BuildCellRangerDataset}
\alias{BuildCellRangerDataset}
\title{BuildCellRangerDataset}
\usage{
BuildCellRangerDataset(data.dir)
}
\description{
Called by \code{\link{LoadCellRangerData}} as a part of \code{\link{CellRangerToASCEND}}. This function reads in the following files:
\itemize{
\item{\strong{barcodes.tsv} - Filtered cellular barcodes recognised by Cell Ranger}
\item{\strong{genes.tsv} - ENSEMBL transcript IDs and their associated gene symbols}
\item{\strong{matrix.mtx} - The expression matrix in Market Matrix format}
}
}
