% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_cellranger.R
\name{BuildExpressionMatrix}
\alias{BuildExpressionMatrix}
\title{BuildExpressionMatrix}
\usage{
BuildExpressionMatrix(sparse.matrix, barcode.obj, gene.obj)
}
\description{
Called by \code{\link{BuildCellRangerDataset}} to do the following:
\enumerate{
\item Convert the loaded \linkS4class{dgTMatrix} to \linkS4class{dgCMatrix}
\item Add cellular barcodes as column labels
\item Add gene names as row labels
}
}
\details{
\linkS4class{dgCMatrix} objects are preferred due to their smaller size.
}
