% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_clustering.R
\name{BuildKeyStat}
\alias{BuildKeyStat}
\title{BuildKeyStat
Generates a bunch of values based on other calculates. Called by RunCORE.}
\usage{
BuildKeyStat(rand.matrix)
}
\description{
BuildKeyStat
Generates a bunch of values based on other calculates. Called by RunCORE.
}
