% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{CalcGeoMeans}
\alias{CalcGeoMeans}
\title{CalcGeoMeans}
\usage{
CalcGeoMeans(x)
}
\arguments{
\item{x}{Counts associated with a cell}
}
\value{
Geometric means associated with each cell
}
\description{
Calculate the geometric mean around a point. Called by 
\code{\link{NormaliseByRLE}} function.
}
