% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcGeoMeansDense}
\alias{calcGeoMeansDense}
\title{Calculate geometric means for dense matrices}
\usage{
calcGeoMeansDense(x)
}
\arguments{
\item{x}{A NumericMatrix}
}
\description{
Calculate geometric means for dense matrices
}
