% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcGeoMeansSparse}
\alias{calcGeoMeansSparse}
\title{Calculate geometric means of sparse matrices}
\usage{
calcGeoMeansSparse(sparseMatrix)
}
\arguments{
\item{sparseMatrix}{Sparse matrix from R}
}
\description{
Calculate geometric means of sparse matrices
}
