% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{CalcNormFactor}
\alias{CalcNormFactor}
\title{CalculateNormFactor}
\usage{
CalcNormFactor(x, geo.means)
}
\arguments{
\item{x}{List of counts associated with a gene.}

\item{geo.means}{Geometric mean associated with a cell as calculated by the
\code{\link{CalcGeoMeans}} function.}
}
\value{
A list of normalisation factors.
}
\description{
Calculate the normalisation factor between all cells. This function is called
by \code{\link{NormaliseByRLE}}.
}
