% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcNormFactorDense}
\alias{calcNormFactorDense}
\title{Calculate size factors for dense matrices}
\usage{
calcNormFactorDense(x, y)
}
\arguments{
\item{x}{matrix}

\item{y}{geometric means}
}
\value{
A numeric vector with sizeFactors for each cell
}
\description{
Calculate size factors for dense matrices
}
