% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcNormFactorSparse}
\alias{calcNormFactorSparse}
\title{Calculate size factors from sparse matrices}
\usage{
calcNormFactorSparse(sparseMatrix, geoMeans)
}
\arguments{
\item{sparseMatrix}{A sparse matrix from R}

\item{geoMeans}{List of size factors}
}
\description{
Calculate size factors from sparse matrices
}
