% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_dimreduction.R
\name{CalcRowVariance}
\alias{CalcRowVariance}
\title{CalcRowVariance}
\usage{
CalcRowVariance(x)
}
\arguments{
\item{x}{A matrix to calculate the row-wise variance of}
}
\value{
A vector of numeric values
}
\description{
Internal function called by RunPCA function
}
