% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\docType{methods}
\name{calculateCV,EMSet-method}
\alias{calculateCV,EMSet-method}
\title{calculateCV}
\usage{
\S4method{calculateCV}{EMSet}(object)
}
\description{
Calculates the Coefficient of Variation (CV) for each gene using normalised
values. Results are stored in the rowData slot.
}
\examples{
# Load example dataset
load("R/sysdata.rda")
x <- data_package$processed_emset

# Calculate CV 
x <- calculateCV(x)

# Show CV results
rowData(x)

}
