% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\docType{methods}
\name{calculateCV}
\alias{calculateCV}
\alias{calculateCV,EMSet-method}
\title{calculateCV}
\usage{
calculateCV(object)

\S4method{calculateCV}{EMSet}(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet}}
}
\value{
An EMSet with calculated CV values.
}
\description{
Calculates the Coefficient of Variation (CV) for each gene using normalised
values. Results are stored in the rowData slot.
}
\examples{
# Load example dataset
x <- ascend::analyzed_set

# Calculate CV 
x <- calculateCV(x)

# Show CV results
rowData(x)

}
