% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\docType{methods}
\name{calculateQC,EMSet-method}
\alias{calculateQC,EMSet-method}
\title{calculateQC}
\usage{
\S4method{calculateQC}{EMSet}(object)
}
\description{
Calculates the following values for quality control:
}
\details{
\describe{
  \item{qc_libsize:}{Total number of counts per cell}
  \item{qc_ngenes}{Number of genes expressed by a cell}
  \item{qc_ncounts}{Total counts per gene} 
  \item{qc_ncells}{Number of cells expressing gene}
  \item{qc_meancounts}{Mean expression of the gene}
  \item{qc_topgeneranking}{Rank in most abundant gene expression}
  \item{qc_pct_total_expression}{Proportion of gene expression to total expression}
}

If users have defined controls, the following metrics are also calculated:
\describe{
  \item{qc_nfeaturecounts}{Number of reads from non-control genes (feature) for each cell}
  \item{qc_pctfeatures}{Percentage of features to total expression}
}

This function is called when changes are made to the count matrix.
}
\examples{
# Load example dataset
object <- ascend::data_package$raw_emset

# Recalculate quality control metrics
object <- calculateQC(object)

}
