% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_clustering.R
\name{CalculateRandIndex}
\alias{CalculateRandIndex}
\title{CalculateRandIndex}
\usage{
CalculateRandIndex(x, adjust = TRUE)
}
\arguments{
\item{x}{Set of clustering results.}

\item{adjust}{TRUE or FALSE.}
}
\value{
A matrix of rand indices based on \code{x}
}
\description{
Calculates rand matrix based on clustering results
}
