% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_cellranger.R
\name{CellRangerToascend}
\alias{CellRangerToascend}
\title{CellRangerToascend}
\usage{
CellRangerToascend(cellranger.path, genome.name)
}
\arguments{
\item{cellranger.path}{This is the path to the folder of data generated by
Cell Ranger's count pipeline.}

\item{genome}{This is the name of the reference genome used by Cell Ranger.}
}
\value{
An \linkS4class{EMSet}
}
\description{
\code{\link{CellRangerToascend}} processes data generated by Cell Ranger's
count pipeline into a \linkS4class{EMSet} object, for use with \pkg{ascend}.
This function assumes the user is using ribosomal and mitochondrial genes as
a control for the experiment.
}
