% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{CheckCellInformation}
\alias{CheckCellInformation}
\title{CheckCellInformation}
\usage{
CheckCellInformation(object, errors)
}
\arguments{
\item{object}{An EMSEt object}

\item{errors}{Vector to store error messages in. If the object is returned
with values - the object will be declared invalid.}
}
\description{
Checks the validity of cell information supplied by the user, to ensure it
matches the information held in the expression matrix.
}
