% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_cellranger.R
\name{CheckCellRangerDirs}
\alias{CheckCellRangerDirs}
\title{CheckCellRangerDirs}
\usage{
CheckCellRangerDirs(target.dir, cellranger.dir)
}
\description{
Called by \code{\link{LoadCellRangerData}}. This function is involved in checks to make sure the Cell Ranger output folder is structured correctly.
}
