% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\name{CheckControls}
\alias{CheckControls}
\title{CheckControls
Check controls if they are present, for issues such as mismatches.}
\usage{
CheckControls(object, errors)
}
\description{
CheckControls
Check controls if they are present, for issues such as mismatches.
}
