% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_utilities.R
\name{CheckData}
\alias{CheckData}
\title{CheckData
Basic function to check if all the items in one list, is present in another 
list}
\usage{
CheckData(object, query)
}
\arguments{
\item{object}{List 1}

\item{query}{List 2}
}
\value{
List of missing objects
}
\description{
CheckData
Basic function to check if all the items in one list, is present in another 
list
}
