% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{CheckGeneInformation}
\alias{CheckGeneInformation}
\title{CheckGeneInformation}
\usage{
CheckGeneInformation(object, errors)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object}

\item{errors}{Vector to store error messages in. If the object is returned
with values - the object will be declared invalid.}
}
\value{
Collected error messages stored in the \code{errors} list
}
\description{
Checks the validity of gene information if supplied by the user. This is to
ensure the gene information in the first column matches the information 
stored in the expression matrix.
}
