% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_clustering.R
\name{ChooseNew}
\alias{ChooseNew}
\title{ChooseNew}
\usage{
ChooseNew(n = NULL, k = NULL)
}
\arguments{
\item{n}{First clustering result.}

\item{k}{Second clustering result.}
}
\description{
Function to calculate rand index. Adapted from function by Steve Horvarth and
Luohua Jiang (UCLA, 2003).
}
