% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getters.R
\docType{methods}
\name{clusterAnalysis,EMSet-method}
\alias{clusterAnalysis,EMSet-method}
\title{clusterAnalysis}
\usage{
\S4method{clusterAnalysis}{EMSet}(x)
}
\description{
Convenience method to retrieve cluster analysis data from the slot.
}
