% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getters.R
\docType{methods}
\name{colInfo,EMSet-method}
\alias{colInfo,EMSet-method}
\title{colInfo}
\usage{
\S4method{colInfo}{EMSet}(x, withDimnames = TRUE)
}
\description{
Convenience method to retrieve cell information.
}
