% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getters.R, R/ascend_setters.R
\docType{methods}
\name{colInfo}
\alias{colInfo}
\alias{colInfo,EMSet-method}
\alias{colInfo<-}
\alias{colInfo<-,EMSet-method}
\title{colInfo}
\usage{
colInfo(x, ...)

\S4method{colInfo}{EMSet}(x, withDimnames = TRUE)

colInfo(x, ...) <- value

\S4method{colInfo}{EMSet}(x) <- value
}
\arguments{
\item{x}{\linkS4class{EMSet}}

\item{...}{...}

\item{withDimnames}{Retrieve rownames of DataFrame Default: TRUE}

\item{value}{DataFrame to store in colInfo slot.}
}
\value{
colInfo DataFrame
}
\description{
Access cell information stored in colInfo.
}
\examples{
# Load EMSet
em_set <- ascend::raw_set

# Retrieve colInfo
column_info <- colInfo(em_set)

# Set colInfo
colInfo(em_set) <- column_info

}
