% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_utilities.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(x, to = c("sce", "seurat", "scone", "EMSet"))
}
\arguments{
\item{x}{Object to convert}

\item{to}{sce, seurat, scone, EMSet}
}
\value{
Object in specified format
}
\description{
Conversion function for working with the `ascend`` package. This conversion
function will convert an EMSet to the following classes: SingleCellExperiment,
Seurat, and SconeExperiment. This function will also convert 
SingleCellExperiment and Seurat objects to an EMSet. 
To convert between Seurat and SingleCellExperiment, it is recommended you use 
the instructions for Seurat [here](https://satijalab.org/seurat/conversion_vignette.html#converting-tofrom-singlecellexperiment).
}
