% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\name{ConvertGeneAnnotation}
\alias{ConvertGeneAnnotation}
\title{ConvertGeneAnnotation}
\usage{
ConvertGeneAnnotation(object, old.annotation, new.annotation)
}
\arguments{
\item{object}{A \linkS4class{EMSet} object.}

\item{old.annotation}{Name of the column containing the current gene annotations.}

\item{new.annotation}{Name of the column you would like to convert the gene 
annotations to.}
}
\value{
An EMSet with the row names in the expression matrix switched to the
values in the new annotation column, and controls converted to their 
corresponding identifier in the new annotation.
}
\description{
Convert gene identifiers used in this \linkS4class{EMSet} to identifiers used 
in another column of a dataframe stored in the GeneAnnotation slot.
}
