% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\docType{methods}
\name{convertGeneID}
\alias{convertGeneID}
\alias{convertGeneID,EMSet-method}
\title{convertGeneID}
\usage{
convertGeneID(object, ..., new.annotation)

\S4method{convertGeneID}{EMSet}(object, ..., new.annotation = NULL)
}
\arguments{
\item{object}{\linkS4class{EMSet}}

\item{...}{...}

\item{new.annotation}{Name of the column where the new set of identifiers 
are stored in rowInfo}
}
\description{
Switches gene identifiers used in the EMSet to the identifiers stored in a 
column of rowInfo.
}
\examples{
# Load data
x <- ascend::raw_set

# Switch identifiers
x <- convertGeneID(x, new.annotation = "ensembl_gene_id")

}
