% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_utilities.R
\name{ConvertMatrix}
\alias{ConvertMatrix}
\title{ConvertMatrix}
\usage{
ConvertMatrix(matrix, format = c("data.frame", "matrix", "sparse.matrix"))
}
\arguments{
\item{matrix}{A data frame, dense matrix or sparse matrix.}

\item{format}{Format to change the matrix to.}
}
\description{
Converts matrix to format of choice. This function gets called by other functions.
}
