% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_scran.R
\name{ConvertScaterToASCEND}
\alias{ConvertScaterToASCEND}
\title{ConvertScaterToASCEND}
\usage{
ConvertScaterToASCEND(SCESet, AEMSet)
}
\arguments{
\item{SCESet}{A \linkS4class{SCESet} from \pkg{scater}}

\item{AEMSet}{An \linkS4class{AEMSet}}
}
\description{
Loads data from a SCATER object to a pre-existing ASCEND object.
}
