% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertSparse}
\alias{convertSparse}
\title{Convert R sparse matrix to RcppArmadillo sparse matrix}
\usage{
convertSparse(mat)
}
\arguments{
\item{mat}{Sparse matrix object from R}
}
\description{
Convert R sparse matrix to RcppArmadillo sparse matrix
}
