% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_scran.R
\name{ConvertToScater}
\alias{ConvertToScater}
\title{ConvertToScater}
\usage{
ConvertToScater(object, control.list = list())
}
\arguments{
\item{object}{An \linkS4class{AEMSet} object}

\item{control.list}{Optional - a named list containing mitochondrial and ribosomal genes}
}
\description{
Convert a \linkS4class{AEMSet} object into a \linkS4class{SCESet} for use with \pkg{scater} and \pkg{scran}.
In order to use this function, you must have mitochondrial and ribosomal genes in your expression data.
}
