% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_scran.R
\name{ConvertToSCE}
\alias{ConvertToSCE}
\title{ConvertToSCE}
\usage{
ConvertToSCE(object, control.list = list())
}
\arguments{
\item{object}{An \linkS4class{EMSet} object.}

\item{control.list}{Optional - a named list containing mitochondrial and ribosomal genes.}
}
\description{
Convert a \linkS4class{EMSet} object into a \linkS4class{SingleCellExperiment} 
for use with \pkg{scater}, \pkg{scran} and other Bioconductor packages.
In order to use this function, you must have mitochondrial and ribosomal 
genes in your expression data.
}
