% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_scran.R
\name{ConvertToSCESet}
\alias{ConvertToSCESet}
\title{ConvertToSCESet}
\usage{
ConvertToSCESet(object, control.list = list())
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object.}

\item{control.list}{Optional - a named list containing mitochondrial and 
ribosomal genes.}
}
\value{
A SCESet
}
\description{
Convert a \code{\linkS4class{EMSet}} object into a \pkg{scater} SCESet for 
use with older versions of \pkg{scater} and \pkg{scran}. In order to use this 
function, you must have mitochondrial and ribosomal genes in your expression 
data.
}
\examples{
\dontrun{
sce_set <- ConvertToSCESet(em.set, 
control.list = list(Mt = mt.genes, Rb = rb.genes))
}
}
