% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\name{DisplayLog}
\alias{DisplayLog}
\title{DisplayLog}
\usage{
DisplayLog(x)
}
\arguments{
\item{x}{An \linkS4class{EMSet}}
}
\value{
A print-out of the log attached to an \code{x}
}
\description{
Print out the log of an \code{\linkS4class{EMSet}}.
}
\examples{
# Load example EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", 
"ExampleClusteredEMSet.rds"))

DisplayLog(EMSet)

}
