% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\docType{methods}
\name{EMSet2SCE,EMSet-method}
\alias{EMSet2SCE,EMSet-method}
\title{EMSet2SCE}
\usage{
\S4method{EMSet2SCE}{EMSet}(x)
}
\arguments{
\item{x}{An EMSet}
}
\value{
A SingleCellExperiment object.
}
\description{
Converts an EMSet to a SingleCellExperiment object. This function preserves
the information stored in the colInfo and rowInfo steps by merging them with
colData and rowData, respectively. Users should extract information in
the log and clusterAnalysis slots prior to conversion.
}
