% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\name{ExcludeControl}
\alias{ExcludeControl}
\title{ExcludeControl}
\usage{
ExcludeControl(object, control.name)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object. It is recommended that you run 
this step after this object has undergone filtering}

\item{control.name}{Name of the control set you want to remove from the dataset}
}
\value{
An \linkS4class{EMSet} without controls in the expression matrix
}
\description{
Removes the specified control from the expression matrix. This should be used
after normalisation, and prior to PCA reduction, clustering and differential
expression analysis.
}
