% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_utilities.R
\name{fileCheck}
\alias{fileCheck}
\title{fileCheck 
This little snippet is designed in a way in which you pair it up with a call 
to process the file. This checks if the file exists.}
\usage{
fileCheck(x)
}
\arguments{
\item{x}{Name of the file you want to check.}
}
\value{
Returns a message if the file doesn't exist.
}
\description{
fileCheck 
This little snippet is designed in a way in which you pair it up with a call 
to process the file. This checks if the file exists.
}
