% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_filtering.R
\name{FilterByCustomControl}
\alias{FilterByCustomControl}
\title{FilterByCustomControl}
\usage{
FilterByCustomControl(control.name, percentage.threshold, object)
}
\arguments{
\item{control.name}{Name of the control group, as used in the named list supplied to the AEMSet object}

\item{percentage.threshold}{Percentage threshold to filter cells by, as a decimal.}

\item{object}{A \linkS4class{AEMSet} object.}
}
\description{
Filter cells in an expression matrix based on the expression levels of a specific control.
This function should be used AFTER the cells have undergone general filtering with the \code{\link{FilterByOutliers}} function.
}
