% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_filtering.R
\name{FilterByExpressedGenesPerCell}
\alias{FilterByExpressedGenesPerCell}
\title{FilterByExpressedGenesPerCell}
\usage{
FilterByExpressedGenesPerCell(object, pct.value)
}
\arguments{
\item{object}{AEMSet (ASCEND) object that has been filtered by \code{\link{FilterByOutliers}} and \code{\link{FilterByCustomControl}}.}

\item{pct.value}{Percentage threshold as a decimal.}
}
\description{
Filtered cells by the proportion of expressed genes in the cell. This step is usually done after the other filtering steps and prior to normalisation.
}
