% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_filtering.R
\name{FilterByOutliers}
\alias{FilterByOutliers}
\title{FilterByOutliers
Automatically filter cells based on expression levels}
\usage{
FilterByOutliers(object, cell.threshold = 3, control.threshold = 3)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object}

\item{cell.threshold}{Mean Absolute Deviation (MAD) value to filter cells by library size. Default: 3}

\item{control.threshold}{Mean Absolute Deviation (MAD) value to filter cells by proportion of control genes. Default: 3}
}
\description{
These values are then used to filter out cells based on the following criteria:
\itemize{
\item{Low overall gene expression}
\item{Low number of expressed genes}
\item{Expression of control genes beyond set threshold}
}
}
\details{
This function then loads the filtered expression matrix into the EMSet object.
}
