% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_filtering.R
\name{filterControl}
\alias{filterControl}
\title{filterControl}
\usage{
filterControl(control_group, total_counts, expression_matrix)
}
\arguments{
\item{control_group}{Name of control group.}

\item{total_counts}{List of total counts of each cell.}

\item{expression_matrix}{An expression matrix.}
}
\value{
Percentage expression of controls per cell.
}
\description{
Called by \code{\link{filterByOutliers}}. This function identifies cells to
remove based on expression levels of control genes.
}
