% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_DESeq.R
\name{func_DESeq}
\alias{func_DESeq}
\title{func_DESeq}
\usage{
func_DESeq(x, expression_matrix = NULL, condition_list = NULL,
  condition.a = NULL, condition.b = NULL, fitType = NULL,
  method = NULL)
}
\arguments{
\item{x}{List of genes}

\item{expression_matrix}{Matrix of counts with a pseudocount added}

\item{condition_list}{Vector of conditions linked to cells}

\item{condition.a}{Condition A to test}

\item{condition.b}{Condition B to test}

\item{fitType}{DESeq argument - fitType to use}

\item{method}{DESeq argument - method to use}
}
\description{
Function called by parallel version of runDESeq.
}
