% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_diffexpression.R
\name{GenerateConditionList}
\alias{GenerateConditionList}
\title{GenerateConditionList}
\usage{
GenerateConditionList(condition.a = NULL, condition.b = NULL,
  barcode.list = NULL)
}
\description{
Automates the generation of a condition 1 vs other clist for feeding into DESeq.
}
