% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\name{GenerateControlMetrics}
\alias{GenerateControlMetrics}
\title{GenerateControlMetrics}
\usage{
GenerateControlMetrics(x, expression.matrix, control.list, total.counts)
}
\arguments{
\item{x}{Name of a control listed in Controls}

\item{expression.matrix}{Transcript counts stored in a sparse matrix object}

\item{control.list}{A list of controls supplied by the user}

\item{total.counts}{List of total counts generated by \code{link{GenerateMetrics}}}
}
\description{
Function called by \code{\link{GenerateMetrics}} to generate metrics for control-related statistics.
}
