% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_cellranger.R
\name{GenerateDefaultControls}
\alias{GenerateDefaultControls}
\title{GenerateDefaultControls}
\usage{
GenerateDefaultControls(gene.obj)
}
\description{
Called by \code{\link{BuildCellRangerDataset}} to retrieve mitochondrial and
ribosomal gene names and list them as controls.
}
