% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\docType{methods}
\name{GenerateMetrics,EMSet-method}
\alias{GenerateMetrics,EMSet-method}
\title{GenerateMetrics}
\usage{
\S4method{GenerateMetrics}{EMSet}(x)
}
\arguments{
\item{x}{An EMSet}
}
\value{
A list of metrics attached to \code{x}
}
\description{
This function generates the following set of values per cell:
\itemize{
\item{\strong{Total Counts}: Total number of expressed transcripts per cell.}
\item{\strong{Total Feature Counts per Cell}: Number of non-control genes expressed per cell.}
\item{\strong{Total Expression}: Total number of expressed transcripts in the dataset.}
\item{\strong{Top Gene List}: Entire list of genes organised from most expressed to least expressed across the entire dataset.}
\item{\strong{Control Transcript Counts}: Total sum of control transcript counts per cell.}
\item{\strong{Percentage Total Counts}: Percentage of transcripts originating from control genes per cell.}
\item{\strong{AverageCounts}: Average transcript count for a gene.}
\item{\strong{GenesPerCell}: Number of unique transcripts expressed by a cell.}
\item{\strong{CellsPerGene}: Number of cells expressing a gene.}
\item{\strong{CountsPerGene}: Total number of transcripts produced by that gene expressed by all cells.}
\item{\strong{Mean Gene Expression}: Mean expression level of a gene across the entire dataset.}
}
}
\details{
This function is called by \code{\link{NewEMSet}} and generates metrics for 
the new expression matrix. This function can also be called independantly, to 
update the metrics for a \code{\linkS4class{EMSet}} object.
}
