% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\name{GenerateMetrics}
\alias{GenerateMetrics}
\title{GenerateMetrics}
\usage{
GenerateMetrics(x)
}
\arguments{
\item{x}{An EMSet}
}
\value{
A list of metrics attached to \code{x}
}
\description{
Generic method for \code{\link{GenerateMetrics}}.
}
