% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_clustering.R
\name{GenerateRandIndexMatrix}
\alias{GenerateRandIndexMatrix}
\title{GenerateRandIndexMatrix}
\usage{
GenerateRandIndexMatrix(cluster.matrix, original.clusters, cluster.list)
}
\description{
Generates columns in KeyStats data frame related to rand index calculations.
}
