% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_methods.R
\name{GetBatchInfo}
\alias{GetBatchInfo}
\title{GetBatchInformation}
\usage{
GetBatchInfo(object)
}
\value{
A list of cell identifiers and their associated batch identifier.
}
\description{
Retrieve batch information from an \linkS4class{AEMSet}.
}
