% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\docType{methods}
\name{GetBatchMatrix,EMSet-method}
\alias{GetBatchMatrix,EMSet-method}
\title{GetBatchMatrix}
\usage{
\S4method{GetBatchMatrix}{EMSet}(object, batch.id)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}

\item{batch.id}{Batch identifier that you would like to retrieve}
}
\value{
Gene/transcript counts for cells in \code{batch.id}
}
\description{
Retrieve a portion of the matrix by batch label
}
\examples{
# Load EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", "ExampleEMSet.rds"))

# Retrieve data frame containing gene information from the object
matrix1 <- GetBatchMatrix(EMSet, 1)

}
