% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\docType{methods}
\name{GetCellInfo,EMSet-method}
\alias{GetCellInfo,EMSet-method}
\title{GetCellInfo}
\usage{
\S4method{GetCellInfo}{EMSet}(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}
}
\value{
A data frame with cell identifiers and associated information.
}
\description{
Retrieve cell information from an \code{\linkS4class{EMSet}}.
}
