% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_getfunctions.R
\name{GetCellInfo}
\alias{GetCellInfo}
\title{GetCellInfo}
\usage{
GetCellInfo(object)
}
\value{
A data frame with cell identifiers and associated information.
}
\description{
Retrieve cell information from an \linkS4class{EMSet}.
}
