% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_methods.R
\name{GetClusters}
\alias{GetClusters}
\title{GetClusters}
\usage{
GetClusters(object)
}
\arguments{
\item{object}{An \linkS4class{AEMSet} that has undergone clustering.}
}
\value{
A list of cell identifiers and their assigned cluster.
}
\description{
GetClusters
}
