% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_clustering.R
\name{getConsecSeq}
\alias{getConsecSeq}
\title{getConsecSeq}
\usage{
getConsecSeq(x, direction = c("forward", "reverse"))
}
\arguments{
\item{x}{Vector of numbers to identify consecutive regions in.}

\item{direction}{Forward or Reverse.}
}
\value{
A list of values that are consecutive
}
\description{
Function that looks for consecutive values in a forward or reverse direction.
Used to calculate stability in the runCORE algorithm.
}
