% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_clustering.R
\name{GetConsecutiveSequence}
\alias{GetConsecutiveSequence}
\title{GetConsecutiveSequence}
\usage{
GetConsecutiveSequence(x, direction = c("forward", "reverse"))
}
\arguments{
\item{x}{Set of values to find the consecutive sequences in}

\item{direction}{Direction to find the consecutive sequences in}
}
\value{
A list of consecutive numbers
}
\description{
Function used to determine consecutive sequences on boundaries.
This is called by FindOptimalResult to determine which clustering results are
consecutive.
}
