% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_getfunctions.R
\name{GetControls}
\alias{GetControls}
\title{GetControls}
\usage{
GetControls(object)
}
\arguments{
\item{object}{An \linkS4class{AEMSet} object.}
}
\value{
This function returns a list of controls defined by the user.
}
\description{
Retrieve list of controls from \linkS4class{AEMSet}.
}
