% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\docType{methods}
\name{GetDistanceMatrix,EMSet-method}
\alias{GetDistanceMatrix,EMSet-method}
\title{GetDistanceMatrix}
\usage{
\S4method{GetDistanceMatrix}{EMSet}(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object.}
}
\value{
This function returns a distance matrix.
}
\description{
Retrieve the distance matrix from an \code{\linkS4class{EMSet}} object that 
has undergone clustering.
}
