% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\docType{methods}
\name{GetExpressionMatrix,EMSet-method}
\alias{GetExpressionMatrix,EMSet-method}
\title{GetExpressionMatrix}
\usage{
\S4method{GetExpressionMatrix}{EMSet}(object, format = c("data.frame",
  "matrix", "sparseMatrix"))
}
\arguments{
\item{object}{A \code{\linkS4class{EMSet}} to retrieve the expression matrix 
from}

\item{format}{Format of the returned matrix - 'data.frame' or 'matrix'}
}
\value{
Returns the expression matrix in the chosen format (data.frame or 
matrix).
}
\description{
Returns a data frame containing the expression matrix from an 
\code{\linkS4class{EMSet}} object.
}
