% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\docType{methods}
\name{GetGeneInfo,EMSet-method}
\alias{GetGeneInfo,EMSet-method}
\title{GetGeneInfo}
\usage{
\S4method{GetGeneInfo}{EMSet}(object)
}
\arguments{
\item{object}{A \code{\linkS4class{EMSet}} object.}
}
\value{
This function returns a data frame containing information about 
the genes in \code{object}.
}
\description{
Retrieve gene information from an \code{\linkS4class{EMSet}}.
}
