% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\docType{methods}
\name{GetHclust,EMSet-method}
\alias{GetHclust,EMSet-method}
\title{GetHclust}
\usage{
\S4method{GetHclust}{EMSet}(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}
}
\value{
This function returns the hclust object stored in \code{object}.
}
\description{
Retrieve the Hclust object from an \code{\linkS4class{EMSet}} object that has 
undergone clustering.
}
