% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\name{GetHclust}
\alias{GetHclust}
\title{GetHclust}
\usage{
GetHclust(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}
}
\value{
This function returns the hclust object stored in \code{object}
}
\description{
Generic for \code{\link{GetHclust}}.
}
\examples{
# Load EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", 
"ExampleClusteredEMSet.rds"))

# Retrieve hclust from the object
hClust <- GetHclust(EMSet)

}
