% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\docType{methods}
\name{GetPCA,EMSet-method}
\alias{GetPCA,EMSet-method}
\title{GetPCA}
\usage{
\S4method{GetPCA}{EMSet}(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}
}
\value{
PCA matrix stored in \code{object}
}
\description{
Retrieve the PCA matrix from an \code{\linkS4class{EMSet}} object that has 
undergone PCA reduction.
}
