% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\name{GetPCA}
\alias{GetPCA}
\title{GetPCA}
\usage{
GetPCA(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}
}
\value{
PCA matrix stored in \code{object}
}
\description{
Generic for \code{\link{GetPCA}}.
}
