% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\docType{methods}
\name{GetRandMatrix,EMSet-method}
\alias{GetRandMatrix,EMSet-method}
\title{GetRandMatrix}
\usage{
\S4method{GetRandMatrix}{EMSet}(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object that has undergone 
clustering}
}
\value{
This function returns a data frame containing cluster stability 
information.
}
\description{
Retrieve the rand matrix used to determine clusters from an 
\code{\linkS4class{EMSet}} object.
}
\examples{
# Load EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", 
"ExampleClusteredEMSet.rds"))

# Retrieve rand matrix from the object
randMatrix <- GetRandMatrix(EMSet)

}
