% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\name{GetRandMatrix}
\alias{GetRandMatrix}
\title{GetRandMatrix}
\usage{
GetRandMatrix(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object that has undergone 
clustering}
}
\value{
This function returns a data frame containing cluster stability 
information.
}
\description{
Generic for \code{\link{GetRandMatrix}}.
}
\examples{
# Load EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", 
"ExampleClusteredEMSet.rds"))

# Retrieve controls from the object
randMatrix <- GetRandMatrix(EMSet)

}
