% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_dimreduction.R
\name{GetReducedDimensions}
\alias{GetReducedDimensions}
\title{GetReducedDimensions}
\usage{
GetReducedDimensions(object, n = 10)
}
\arguments{
\item{object}{An \linkS4class{AEMSet} object that has undergone PCA reduction.}

\item{n}{The number of PC dimensions you would like to select. Refer to vignette on how to select this value.}
}
\description{
GetReducedDimensions
}
