% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_cellranger.R
\name{IdentifyBatches}
\alias{IdentifyBatches}
\title{IdentifyBatches}
\usage{
IdentifyBatches(barcode.obj)
}
\description{
Called by \code{\link{BuildCellRangerDataset}} to parse batch labels from barcodes.tsv file. The result is a named list featuring cell barcodes as the key and batch ID as the value.
}
