% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_setup.R
\name{initial_setup}
\alias{initial_setup}
\title{Set up the environment}
\usage{
initial_setup(install_tinytex = FALSE)
}
\description{
Installs the \code{bibtex} package from Github and installs the \code{tinytex} utility
}
