% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_utilities.R
\name{joinPaths}
\alias{joinPaths}
\title{joinPaths}
\usage{
joinPaths(x)
}
\arguments{
\item{x}{A list of substrings to combine into a full path.}
}
\description{
Convenience function - joins a list of strings to form a full path. 
Disregards "/" at the end of a substring to ensure final full path is not
malformed.
}
\examples{
sub_paths <- c("Path", "To/", "Me")
full_path <- joinPaths(sub_paths)
print(full_path)

}
