% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_cellranger.R
\name{loadCellRanger}
\alias{loadCellRanger}
\title{loadCellRanger}
\usage{
loadCellRanger(x)
}
\arguments{
\item{x}{Path to Cell Ranger filtered output (eg. outs/filtered_gene_bc_matrices_mex/GRCh38p7)}
}
\value{
A list containing the following objects: gene_info, cell_info, counts
}
\description{
Automatically loads and prepares data from a folder containing Cell Ranger
output. Filtered data should be used as raw data is too large for most 
desktop systems.
}
