% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_cellranger.R
\name{LoadCellRangerData}
\alias{LoadCellRangerData}
\title{LoadCellRangerData}
\usage{
LoadCellRangerData(working.dir, genome.name)
}
\description{
Called by \code{\link{CellRangerToascend}}. This function detects which folder
to use in the 'outs' directory of the Cell Ranger output folder.
}
