% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_utilities.R
\name{LoadMatrix}
\alias{LoadMatrix}
\title{LoadMatrix}
\usage{
LoadMatrix(x)
}
\arguments{
\item{x}{A dense matrix, data frame or AEMSet.}
}
\description{
Converts a dense matrix or data frame into a sparse matrix.
}
