% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_utilities.R
\name{LoadSparseMatrix}
\alias{LoadSparseMatrix}
\title{LoadSparseMatrix}
\usage{
LoadSparseMatrix(dataframe)
}
\description{
Converts a dataframe to a sparse matrix, for storage in a \linkS4class{AEMSet} object.
}
