% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_utilities.R
\name{MakeGeneCondition}
\alias{MakeGeneCondition}
\title{GetGeneCondition}
\usage{
MakeGeneCondition(object, gene)
}
\arguments{
\item{object}{An \linkS4class{AEMSet}}

\item{gene}{Gene or transcript ID to label}
}
\description{
Generates a condition list of cell identifiers and if they are expressing the specified gene.
}
