% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_utilities.R
\name{mergeDF}
\alias{mergeDF}
\title{mergeDF}
\usage{
mergeDF(x, y, z)
}
\arguments{
\item{x}{Old DataFrame}

\item{y}{Replacement DataFrame}

\item{z}{Name of column(s) to merge DataFrames by}
}
\value{
A merged DataFrame
}
\description{
Convenience function to merge data frames and overwrite pre-existing
columns with the new values
}
