% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_utilities.R
\name{MergeExprsMtx}
\alias{MergeExprsMtx}
\title{MergeExprsMtx}
\usage{
MergeExprsMtx(x)
}
\arguments{
\item{x}{A list of data frames containing expression matrices.}
}
\description{
Merge a list of expression matrices in data frame format into one large data frame.
}
