% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\name{NewAEMSet}
\alias{NewAEMSet}
\title{NewAEMSet}
\usage{
NewAEMSet(ExpressionMatrix = NULL, GeneAnnotation = NULL,
  BatchInformation = NULL, Controls = list(Mt = list(), Rb = list()))
}
\arguments{
\item{ExpressionMatrix}{An expression matrix in data.frame, dgCMatrix (sparse) or matrix format. Rows should represent a transcript and its counts, while columns should represent individual cells. This is usually the end point for Single Cell RNA-Seq pipelines such as Cell Ranger and DropSeq.}

\item{GeneAnnotation}{A data frame containing gene identifiers used in the expression matrix. Other columns can be used to represent corresponding identifiers in other formats (such as ENSEMBL transcript IDs). This is an optional field, and is best used if you need to convert between identifiers.}

\item{BatchInformation}{A named list of cell identifiers (usually barcodes) and an integer representing which batch they belong to. This is an optional field, and it best used for experiments that contain data from multiple samples.}

\item{Controls}{A named list of controls, eg. mitochondrial genes, ribosomal genes and ERCC spike-ins. These genes must be specified using the identifier used in the expression matrix. This is required.}

\item{Conditions}{A named list of conditions for use in differential expression. Supplying this information is optional.}
}
\value{
This function generates an object belonging to the \linkS4class{AEMSet}.
}
\description{
\code{\link{NewAEMSet}} generates a \linkS4class{AEMSet} object for use with the ASCEND package. This object contains an expression matrix, associated metadata, downstream analysis and a log documenting the actions taken to generate this object.
}
\seealso{
\linkS4class{AEMSet}
}
