% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{newEMSet}
\alias{newEMSet}
\title{newEMSet}
\usage{
newEMSet(assays = NULL, colInfo = NULL, rowInfo = NULL, colData = NULL,
  rowData = NULL, controls = NULL)
}
\arguments{
\item{assays}{A list of things}

\item{colInfo}{A Dataframe}

\item{rowInfo}{A Dataframe}

\item{colData}{A Dataframe}

\item{rowData}{A Dataframe}

\item{controls}{A named list}
}
\value{
An \linkS4class{EMSet}.
}
\description{
Creates a new EMSet.
}
