% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_normalisation.R
\name{NormaliseBatches}
\alias{NormaliseBatches}
\title{NormaliseBatches}
\usage{
NormaliseBatches(object)
}
\arguments{
\item{object}{An \linkS4class{AEMSet} object.}
}
\description{
Normalise counts to remove batch effects. This normalisation method is for
experiments where data from batches of different samples are combined without
undergoing library equalisation.
}
\details{
This step must be done prior to any filtering and normalisation between cells.
}
