% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{normaliseByRLE}
\alias{normaliseByRLE}
\title{normaliseByRLE}
\usage{
normaliseByRLE(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} set that has undergone filtering. 
Please ensure spike-ins have been removed before using this function.}
}
\value{
An \code{\linkS4class{EMSet}} with normalised expression values.
}
\description{
Normalisation of expression between cells, by scaling to relative log 
expression (RLE). Only counts that are greater than zero are considered in 
this normalisation method.
}
\examples{
# Load example EMSet
em_set <- ascend::raw_set

# Normalise batches
norm_set <- normaliseByRLE(em_set)

}
