% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{NormaliseLibSize}
\alias{NormaliseLibSize}
\title{NormaliseLibSize}
\usage{
NormaliseLibSize(object)
}
\arguments{
\item{object}{A \linkS4class{EMSet} object. Please remove spike-ins from the expression matrix before normalising.}
}
\description{
Normalise library sizes by scaling.
}
