% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_normalisation.R
\name{NormWithinBatch}
\alias{NormWithinBatch}
\title{NormWithinBatch}
\usage{
NormWithinBatch(batch.id, expression.matrix = NULL, cell.info = NULL)
}
\description{
Called by NormaliseBatches. Performs normalisation within a batch.
}
