% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotBatchNormQC}
\alias{plotBatchNormQC}
\title{plotBatchNormQC}
\usage{
plotBatchNormQC(raw_object = NULL, norm_object = NULL)
}
\arguments{
\item{raw_object}{Dataset prior to batch normalisation}

\item{norm_object}{Dataset after batch normalisation}
}
\value{
A boxplot created with ggplot2
}
\description{
Creates a boxplot that depicts the library sizes of each batch in a dataset.
This function is for the comparison of pre- and post- batch-normalised 
datasets that have been loaded into an EMSet. Cells found in these datasets 
must have the same identifiers.
}
