% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotConsensusBars}
\alias{PlotConsensusBars}
\title{PlotConsensusBars}
\usage{
PlotConsensusBars(dendro, colors, rowLabels = NULL, rowWidths = NULL,
  rowText = NULL, rowTextAlignment = c("left", "center", "right"),
  rowTextIgnore = NULL, textPositions = NULL, addTextGuide = TRUE,
  cex.rowLabels = 1, cex.rowText = 0.8, ...)
}
\description{
Adapted from plotColorUnderTree found in the \pkg{WGCNA} package.
}
