% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotControlHist}
\alias{plotControlHist}
\title{plotControlHist}
\usage{
plotControlHist(object, control = NULL)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object.}

\item{control}{Name of control group to plot.}
}
\value{
A ggplot2 glob containing the histogram.
}
\description{
Generates a histogram of percentage of control per samples.
}
