% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotControlPctPerSample}
\alias{plotControlPctPerSample}
\title{plotControlPctPerSample}
\usage{
plotControlPctPerSample(object, control = NULL)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object.}

\item{control.name}{Name of the control you would like to plot.}
}
\value{
A ggplot2 object containing a violin-beeswarm plot
}
\description{
Generates a violin/beeswarm plot of percentage of control expression per 
sample. Called by \code{\link[ascend:plotGeneralQC]{plotGeneralQC}}.
}
\examples{
\dontrun{
control_violin <- plotControlPctPerSample(EMSet, control = "Mt")
}
}
