% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotControlPercentagesPerSample}
\alias{PlotControlPercentagesPerSample}
\title{PlotControlPercentagesPerSample}
\usage{
PlotControlPercentagesPerSample(object, control.name)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}

\item{control.name}{Name of the control you would like to plot}
}
\description{
Generates a violin/beeswarm plot of percentage of control expression per 
sample. Called by \code{\link{PlotGeneralQC}}.
}
