% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotDEVolcano}
\alias{PlotDEVolcano}
\title{PlotDEVolcano}
\usage{
PlotDEVolcano(de.results, threshold = 0.005, labels = FALSE)
}
\arguments{
\item{de.results}{Differential expression results generated by \code{\link{RunDifferentialExpression}} or \link{\link{RunClusterDiffExpression}}}

\item{threshold}{Quantile threshold to determine significant results by}

\item{labels}{Display names of significant results on plot}
}
\description{
Produces a volcano plot featuring differential expression results.
}
