% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotLibrarySizesPerSample}
\alias{PlotLibrarySizesPerSample}
\title{PlotLibrarySizesPerSample}
\usage{
PlotLibrarySizesPerSample(object)
}
\arguments{
\item{object}{A \linkS4class{EMSet}}
}
\description{
Generates a violin-beeswarm plot of library sizes per sample.
}
