% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotLibsizeBarplot}
\alias{plotLibsizeBarplot}
\title{plotLibsizeBarplot}
\usage{
plotLibsizeBarplot(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object.}
}
\value{
A ggplot2 glob containing the barplot.
}
\description{
Generates a barplot of each cell's library size, and arranges them in 
descending order.
}
\examples{
\dontrun{
libsize_barplot <- plotLibsizeBarplot(EMSet)
}
}
