% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotLibsizeBoxplot}
\alias{plotLibsizeBoxplot}
\title{plotLibsizeBoxplot}
\usage{
plotLibsizeBoxplot(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet}.}
}
\value{
A ggplot2 glob containing the barplot.
}
\description{
Generates a barplot of each cell's library size, and arranges them in 
descending order.
}
\examples{
# Load EMSet
EMSet <- ascend::raw_set

# Plot libsize boxplot
libsize_barplot <- plotLibsizeBoxplot(EMSet)

}
