% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotLibsizeHist}
\alias{plotLibsizeHist}
\title{plotLibsizeHist}
\usage{
plotLibsizeHist(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet}.}
}
\value{
A ggplot2 glob containing the histogram.
}
\description{
Generates a histogram of all cell library sizes.
}
\examples{
# Load EMSet
EMSet <- ascend::raw_set

# Plot libsize histograms
libsize_hist <- plotLibsizeHist(EMSet)

}
