% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotLibsizePerSample}
\alias{plotLibsizePerSample}
\title{plotLibsizePerSample}
\usage{
plotLibsizePerSample(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object.}
}
\value{
A ggplot2 object containing a violin-beeswarm plot
}
\description{
Generates a violin-beeswarm plot of library sizes per sample.
}
