% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotNormalisationQC}
\alias{PlotNormalisationQC}
\title{PlotNormalisationQC}
\usage{
PlotNormalisationQC(Original = NULL, Normalised = NULL)
}
\arguments{
\item{Original}{An un-normalised \linkS4class{AEMSet}}

\item{Normalised}{A normalised \linkS4class{AEMSet}}
}
\description{
Generates a series of plots comparing an un-normalised and a normalised \linkS4class{AEMSet}.
}
