% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotNormQC}
\alias{plotNormQC}
\title{plotNormQC}
\usage{
plotNormQC(object, gene_list = list())
}
\arguments{
\item{object}{A normalised \code{\linkS4class{EMSet}}.}

\item{gene_list}{OPTIONAL: A list of genes to plot expression levels for. 
If not defined, \pkg{ascend} will select GAPDH and MALAT1, or choose two 
genes at random.}
}
\value{
A list containing the following plots:
\itemize{
\item{Library size histograms.}
\item{Scatter boxplots for selected genes.}
\item{Scatter boxplots for each gene.}
}
}
\description{
Generates a series of plots comparing an un-normalised and a normalised 
\code{\linkS4class{EMSet}}.
}
\examples{
# Load normalised EMSet
EMSet <- ascend::analyzed_set

# Plot normalisation
norm_qc <- plotNormQC(EMSet, gene_list = c("GAPDH", "MALAT1"))

}
