% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotOrderedColors}
\alias{PlotOrderedColors}
\title{PlotOrderedColors}
\usage{
PlotOrderedColors(order, colors, rowLabels = NULL, rowWidths = NULL,
  rowText = NULL, rowTextAlignment = c("left", "center", "right"),
  rowTextIgnore = NULL, textPositions = NULL, addTextGuide = TRUE,
  cex.rowLabels = 1, cex.rowText = 0.8, startAt = 0, ...)
}
\description{
Adapted from plotOrderedColors found in the \pkg{WGCNA} package.
}
