% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotPCA}
\alias{PlotPCA}
\title{PlotPCA}
\usage{
PlotPCA(object, dim1 = 1, dim2 = 2, condition = NULL)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object that has undergone PCA.}

\item{dim1}{Principal component to plot on the x-axis.}

\item{dim2}{Principal component to plot on the y-axis.}

\item{condition}{(Optional) Name of the column in CellIdentifiers that 
describe a set of conditions you would like to colour cells by.}
}
\value{
A ggplot glob that contains a scatter plot.
}
\description{
Plot two principal components (PCs) on a scatter plot. This plot corresponds
more closely to the distance between points and therefore is good to use
to review the effectiveness of clustering by the CORE algorithm.
}
\examples{
# Load clustered EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", 
"ExampleClusteredEMSet.rds"))

# Load ggplot2 to add elements to PCA plot
library(ggplot2)

# Plot PCA and add title to it
my_pca_plot <- PlotPCA(EMSet, dim1 = 1, dim2 = 2,  condition = "cluster")  
my_pca_plot <- my_pca_plot + ggtitle("My PCA Plot")

}
