% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotPCAVariance}
\alias{plotPCAVariance}
\title{plotPCAVariance}
\usage{
plotPCAVariance(object, n = 50)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has undergone PCA.}

\item{n}{Number of PCs to view on the plot.}
}
\value{
A ggplot2 glob that contains a scatter qplot.
}
\description{
Generates a scree plot of PCs generated by 
\code{\link[ascend:runPCA]{runPCA}}. 
The principal components (PCs) are sorted from largest percent variance to 
the smallest percent variance. This plot is for determining the optimal 
number of PCs to retain for further analysis.
}
\examples{
# Load EMSet
em_set <- ascend::analyzed_set

# Plot PCA variance
pca_variance_plot <- plotPCAVariance(em_set, n = 10)

}
