% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotStability}
\alias{PlotStability}
\title{PlotStability}
\usage{
PlotStability(object)
}
\arguments{
\item{object}{An \linkS4class(AEMSet) object that has undergone clustering.}
}
\description{
Plots
}
