% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotStabilityDendro}
\alias{PlotStabilityDendro}
\title{PlotStabilityDendro}
\usage{
PlotStabilityDendro(object)
}
\arguments{
\item{object}{A \linkS4class{EMSet} that has undergone clustering.}
}
\description{
PlotStabilityDendro
}
