% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotTopGenesBoxplot}
\alias{plotTopGenesBoxplot}
\title{plotTopGenesBoxplot}
\usage{
plotTopGenesBoxplot(object, n = 20, controls = TRUE)
}
\arguments{
\item{object}{A EMSet object.}

\item{n}{Number of genes to be plotted.}

\item{controls}{Include control genes in plot (Default: TRUE).}
}
\value{
A ggplot glob containing a box scatter plot that represents the 
expression of the most highly expressed genes.
}
\description{
Generates a boxplot using \link[ggplot2]{geom_boxplot} of the most expressed 
genes in the dataset, in a range defined by the user.
}
\examples{
# Load example EMSet
}
