% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotTSNE}
\alias{PlotTSNE}
\title{PlotTSNE}
\usage{
PlotTSNE(object, PCA = TRUE, condition = NULL, seed = 0,
  perplexity = 30, theta = 0.5)
}
\arguments{
\item{object}{An \linkS4class{EMSet}.}

\item{PCA}{Set to FALSE to not use PCA-reduced values}

\item{condition}{(Optional) Name of the column in CellIdentifiers that describe a set of conditions you would like to colour cells by}

\item{seed}{(Optional) Set to a specific value for reproducible TSNE plots}

\item{perplexity}{(Optional) Numeric; perplexity parameter}

\item{theta}{(Optional) Numeric; Speed/accuracy trade-off (increase for less accuracy)}
}
\value{
A ggplot glob that contains a scatter plot
}
\description{
Generates a 2D TSNE plot. As this function requires the reduction of the
expression matrix by TSNE, it takes longer to run in comparison to the other
reduced dimension plots.
}
\details{
If you just require a TSNE matrix, you can generate this with the
\code{\link{RunTSNE}} function.
}
\examples{
\dontrun{
library(ggplot2)
my_tsne_plot <- PlotTSNE(em.set, PCA = TRUE, condition = "cluster", seed = 1)  
my_tsne_plot <- my_tsne_plot + ggtitle("My TSNE Plot")
}
}
