% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotUMAP}
\alias{plotUMAP}
\title{plotUMAP}
\usage{
plotUMAP(object, Dim1 = 1, Dim2 = 2, group = NULL, density = FALSE)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object that hasv values 
calculated by UMAP in the reducedDim slot.}

\item{Dim1}{Dimension to plot on the x-axis.}

\item{Dim2}{Dimension to plot on the y-axis.}

\item{group}{(Optional) Name of the column in colInfo that 
describe a set of conditions you would like to colour cells by.}

\item{density}{Vary alpha by density (Default: FALSE)}
}
\value{
A ggplot glob that contains a scatter plot.
}
\description{
Generates a 2D UMAP plot using a TSNE matrix generated by the 
\code{\link[ascend:runUMAP]{runUMAP}} function.
}
\examples{
\dontrun{
# Dimensionality reduction with t-SNE
EMSet <- runUMAP(EMSet, method = "naive", config = umap.config)

# Plot with t-SNE 
tsne_plot <- plotUMAP(EMSet, Dim1 = 1, Dim2 = 2, group = "cluster", density = FALSE)

}
}
