% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotVariableGenes}
\alias{plotVariableGenes}
\title{plotVariableGenes}
\usage{
plotVariableGenes(object, ngenes = NULL, label.size = 3,
  point.size = 0.5, check.overlap = FALSE)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has had CV values calculated}

\item{ngenes}{Select n most variable genes to plot (Optional)}

\item{label.size}{Size of gene labels}

\item{point.size}{Size of scatter points}

\item{check.overlap}{Hide overlapping labels (Default: FALSE)}
}
\value{
A scatter plot rendered by ggplot2's geom_point function.
}
\description{
Generates a scatter plot to aid in the detection of variable genes. Scatter
plot depicts Correlation of Variance (CV) vs log10(mean gene expression). 
Please use the \code{\link[ascend:calculateCV]{calculateCV}} 
function before using this function.
}
\examples{
\dontrun{
em_set <- calculateCV(em_set)
variable_gene_plot <- plotVariableGenes(em_set, ngenes = 1500, 
label.size = 3, point.size = 0.5, check.overlap = TRUE)
}

}
