% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sftp_file.R
\name{post_sftp_file}
\alias{post_sftp_file}
\title{Post a file to a directory on an FTP/SFTP server}
\usage{
post_sftp_file(
  protocol = "sftp",
  server = "sftp.ascend-innovations.com",
  user = "hospice-shared",
  pwd = Sys.getenv("OHI_SFTP_PWD"),
  sftp_path = "/uploads/BrightView_data",
  file = "active_clients_cls Centerville 1-21-2021.csv"
)
}
\arguments{
\item{protocol}{The protocol of the \code{server} - as a character string (typically "ftp" or "sftp")}

\item{server}{The address of the \code{server} - as a character string}

\item{user}{The user name - as a character string}

\item{pwd}{The password - as a character string}

\item{sftp_path}{The path to the directory on the \code{server} (if exists) where the file will be posted}

\item{file}{The path to the file to be posted to the \code{server}}
}
\description{
Post a file to a directory on an FTP/SFTP server
}
