% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_diffexpression.R
\name{PrepareCountData}
\alias{PrepareCountData}
\title{PrepareCountData}
\usage{
PrepareCountData(object, cells, ngenes)
}
\arguments{
\item{object}{An \linkS4class{EMSet} to perform differential expression on}

\item{cells}{List of cells to extract from the \linkS4class{EMSet}}

\item{ngenes}{Number of cells to extract from the \linkS4class{EMSet}}
}
\value{
A list of chunks of the expression matrix
}
\description{
Called by \code{\link{RunDiffExpression}}. This chunks up the expression 
matrix to feed into \pkg{DESeq}.
}
