% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_diffexpression.R
\name{ProcessDEResults}
\alias{ProcessDEResults}
\title{ProcessDEREsults}
\usage{
ProcessDEResults(output.list)
}
\arguments{
\item{output.list}{List of DESeq resuilts to process and compile}
}
\value{
One data frame containing DESeq results for all genes
}
\description{
Called by \code{\link{RunDiffExpression}}. Compiles the resultant data into 
data frames and converts the results to absolute fold change.
}
