% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_cellranger.R
\name{ProcessGeneReference}
\alias{ProcessGeneReference}
\title{ProcessGeneReference}
\usage{
ProcessGeneReference(gene.obj)
}
\description{
Called by \code{\link{BuildCellRangerDataset}} to parse genes.tsv into a data
frame.
}
