% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getters.R
\docType{methods}
\name{progressLog,EMSet-method}
\alias{progressLog,EMSet-method}
\title{progressLog}
\usage{
\S4method{progressLog}{EMSet}(x)
}
\description{
Covenience method to retrieve the progress log.
}
