% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_setters.R, R/ascend_getters.R
\docType{methods}
\name{progressLog<-}
\alias{progressLog<-}
\alias{progressLog<-,EMSet-method}
\alias{progressLog}
\alias{progressLog,EMSet-method}
\title{progressLog}
\usage{
progressLog(x, ...) <- value

\S4method{progressLog}{EMSet}(x) <- value

progressLog(x, ...)

\S4method{progressLog}{EMSet}(x)
}
\arguments{
\item{x}{\linkS4class{EMSet}}

\item{...}{...}

\item{value}{List to store in the progressLog slot}
}
\value{
A list of information.
}
\description{
Access log stored in the \linkS4class{EMSet}.
}
\examples{
# Load EMSet
em_set <- ascend::analyzed_set

# Retrieve progress log
log <- progressLog(em_set)

# Set progressLog
progressLog(em_set) <- log

}
