% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw}
\alias{raw}
\title{Retinal Ganglion Cells (100 THY-1 positive, 100 THY-1 negative)}
\format{A list consisting of three objects:
\describe{
  \item{counts}{Count matrix stored in sparseMatrix format}
  \item{col_info}{DataFrame containing cell barcode and batch information}
  \item{row_info}{DataFrame containing gene names and corresponding ENSEMBL 
  gene identifiers}
  }}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6108/}
}
\usage{
raw
}
\description{
A dataset containing a count matrix, cell-related metadata and gene-
related metdata.
}
\keyword{datasets}
