% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_set}
\alias{raw_set}
\title{Retinal Ganglion Cells (100 THY-1 positive, 100 THY-1 negative)}
\format{A list consisting of three objects:
\describe{
  \item{counts}{Count matrix stored in sparseMatrix format}
  \item{colInfo}{DataFrame containing cell barcode and batch information}
  \item{rowInfo}{DataFrame containing gene names and corresponding ENSEMBL 
  gene identifiers}
  \item{controls}{A list containing mitochondrial-associated transcripts and
  ribosomal-associated transcripts}
  }}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-6108/}
}
\usage{
raw_set
}
\description{
An EMSet containing the data from "raw" loaded into an EMSet.
}
\keyword{datasets}
