% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_clean.R
\name{read_clean}
\alias{read_clean}
\title{Read and clean data extracted from Brightview}
\usage{
read_clean(
  .data,
  .keys,
  .name = "API",
  .type = "SUD",
  .payer = "PAYER",
  .start = "aDATE",
  .end = "aDATE",
  .description = "DXNUM_EX",
  .id = "ENCOUNTER_ID",
  .code = "DXNUM",
  .category = "",
  .other = list("bvDATE", "bvID")
)
}
\arguments{
\item{.data}{A \code{data.frame}-class object}

\item{.keys}{Character vector of column names to use as keys for sorting and merging}

\item{.name}{Character vector or list of column names containing the patient's first and last name}

\item{.payer}{Character string indicating which column from \code{.data} contains payer information}

\item{.start}{Character string indicating which column from \code{.data} contains the start date of the encounter}

\item{.end}{Character string indicating which column from \code{.data} contains the end date of the encounter}

\item{.description}{Character string indicating which column from \code{.data} contains the description of the encounter}

\item{.other}{List of character string indicating which column(s) from \code{.data} contain other information about the encounter}

\item{.group}{Character string indicating which type of encounter to use for naming}
}
\value{

}
\description{
Read and clean data extracted from Brightview
}
