% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_dimreduction.R
\name{ReduceDimensions}
\alias{ReduceDimensions}
\title{ReduceDimensions}
\usage{
ReduceDimensions(object, n = 10)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object that has undergone PCA reduction.}

\item{n}{The number of PC dimensions you would like to select. Refer to
vignette on how to select this value. Default: 10.}
}
\value{
An \code{\linkS4class{EMSet}} with a PCA matrix of dimensions ncells by 
ndimensions.
}
\description{
ReduceDimensions
}
\examples{
# Load PCA-reduced EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", "ExampleClusteredEMSet.rds"))

# Reduce to 10 PCs
ReducedEMSet <- ReduceDimensions(EMSet, n = 10)
}
