% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_setters.R, R/ascend_getters.R
\docType{methods}
\name{regcounts<-}
\alias{regcounts<-}
\alias{regcounts<-,EMSet-method}
\alias{regcounts}
\alias{regcounts,EMSet-method}
\title{regcounts}
\usage{
regcounts(x, ...) <- value

\S4method{regcounts}{EMSet}(x) <- value

regcounts(x, ...)

\S4method{regcounts}{EMSet}(x)
}
\arguments{
\item{x}{\linkS4class{EMSet}}

\item{...}{...}

\item{value}{A matrix with confounding factors regressed out}
}
\value{
Regressed count matrix
}
\description{
Convenience method to access regressed counts from an \linkS4class{EMSet}.
}
