% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_confoundingfactors.R
\name{RegressConfoundingFactors}
\alias{RegressConfoundingFactors}
\title{RegressConfoundingFactors}
\usage{
RegressConfoundingFactors(object, candidate.genes = c())
}
\arguments{
\item{object}{Expression matrix in data frame format. You may load
it from a EMSet object with the GetExpressionMatrix function.}

\item{candidate.genes}{A list of genes you wish to regress from the dataset.
Refer to the vignette on how to choose genes for regression.}
}
\value{
An \code{\linkS4class{EMSet}} with confounding factors regressed from the
expression values.
}
\description{
This function generates a scaled regression matrix based on candidate genes
supplied by the user. This function is best used after normalisation.
}
\examples{
\dontrun{
# Load EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", "ExampleEMSet.rds"))

# Define genes to regress
confoundingGenes <- c("CDK4","CCND1","NOC2L","ATAD3C", "CCNL2")
regressedSet <- RegressConfoundingFactors(EMSet, 
candidate.genes = confoundingGenes)
}
}
