% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_confoundingfactors.R
\name{RegressGene}
\alias{RegressGene}
\title{RegressGene}
\usage{
RegressGene(gene = NULL, covariate.matrix = NULL,
  expression.matrix = NULL)
}
\arguments{
\item{gene}{Gene to regress residuals from.}

\item{covariate.matrix}{Covariate matrix generated by RemoveConfoundingFactors.}

\item{expression.matrix}{A transposed expression matrix generated by RemoveConfoundingFactors.}
}
\description{
Called by RemoveConfoundingFactors. This step extracts residuals from the supplied gene.
}
