% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\docType{methods}
\name{ReplaceCellInfo,EMSet-method}
\alias{ReplaceCellInfo,EMSet-method}
\title{ReplaceCellInfo}
\usage{
\S4method{ReplaceCellInfo}{EMSet}(object, cell.info)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object.}

\item{cell.info}{A data frame containing cell information. The first column 
must contain cell identifiers that match the cell identifiers used in the 
expression matrix. The second column, while optional - should contain batch 
information.}
}
\value{
An \linkS4class{EMSet} with updated cell information.
}
\description{
Can be called by the user or by a filtering function. Updates Cell 
Information in a \linkS4class{EMSet} object, replacing the old data frame 
with a new one.
}
\examples{
# Load an EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", "ExampleEMSet.rds"))

# Retrieve "Cell Information" dataframe from EMSet
old_cell_info <- GetCellInfo(EMSet)

# Update cell labels
new_cell_info <- old_cell_info
new_cell_info$cells <- TRUE

# Replace Cell Information
updatedEMSet <- ReplaceCellInfo(EMSet, new_cell_info)

}
