% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{ReplaceExpressionMatrix}
\alias{ReplaceExpressionMatrix}
\title{ReplaceExpressionMatrix}
\usage{
ReplaceExpressionMatrix(object, expression.matrix)
}
\description{
Definition for \code{\link{ReplaceExpressionMatrix}} function.
}
