% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_clustering.R
\name{RetrieveCluster}
\alias{RetrieveCluster}
\title{RetrieveCluster}
\usage{
RetrieveCluster(height, hclust.obj = NULL, distance.matrix = NULL)
}
\description{
Generates clusters using dynamicTreeCut's unsupervised cut setting.
}
