% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getters.R
\docType{methods}
\name{rowInfo,EMSet-method}
\alias{rowInfo,EMSet-method}
\title{rowInfo}
\usage{
\S4method{rowInfo}{EMSet}(x, withDimnames = TRUE)
}
\description{
Convenience method to retrieve gene information.
}
