% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_diffexpression.R
\name{RunClusterDiffExpression}
\alias{RunClusterDiffExpression}
\title{RunClusterDiffExpression}
\usage{
RunClusterDiffExpression(object)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object that has undergone clustering with the \link{\code{FindOptimalClusters}} function.}
}
\description{
Compare the differential expression of genes in each cluster versus other clusters.
}
