% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_clustering.R
\name{RunCORE}
\alias{RunCORE}
\title{RunCORE}
\usage{
RunCORE(object, conservative = TRUE, windows = seq(0.025:1, by = 0.025),
  remove_outlier = FALSE)
}
\arguments{
\item{object}{An EMSet object that has undergone PCA reduction.}

\item{conservative}{Use conservative (more stable) clustering result (TRUE or FALSE). Default: TRUE}

\item{windows}{Range to perform cuts on the dendrogram Default: seq(0.025:1, by=0.025).}

\item{remove_outlier}{Remove cells that weren't assigned a cluster with dynamicTreeCut. This is indicative of outlier cells within the sample. Default: TRUE}
}
\description{
This function determines the optimal number of clusters for a dataset.
This function first generates a distance matrix and a hclust object, and then cuts the tree at different heights.
This will return an EMSet with the following objects:
\describe{
    \item{DistanceMatrix}{A distance matrix}
    \item{Hclust}{A hclust object}
    \item{PutativeClusters}{Cluster identities generated by dynamicTreeCut}
    \item{ClusteringMatrix}{A matrix containing a cluster identities from cutting at 40 different heights}
    \item{Clusters}{Optimum cluster identities for each cell}
    \item{NumberOfClusters}{Number of clusters}
    \item{OptimalTreeHeight}{Optimal tree height used to generate cluster identities}
    \item{KeyStats}{A dataframe containing information on each generated clustering result, that is used to determine the optimal cluster number.}
    \item{RandMatrix}{Rand matrix used to determine optimal number of clusters}
}
}
