% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_diffexpression.R
\name{RunDESeq}
\alias{RunDESeq}
\title{RunDESeq}
\usage{
RunDESeq(data, condition.list = list(), condition.a = NULL,
  condition.b = NULL, fitType = NULL, method = NULL)
}
\description{
RunDESeq
}
