% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_DESeq2.R
\name{runDESeq2}
\alias{runDESeq2}
\title{runDESeq2}
\usage{
runDESeq2(object, group = NULL, condition.a = NULL, condition.b = NULL,
  fitType = c("parametric", "local", "mean"), ngenes = NULL)
}
\arguments{
\item{object}{An \linkS4class{EMSet}}

\item{group}{Column in colInfo that contains a set of conditions you want to test}

\item{condition.a}{Condition(s) in the group column that you want to test for.}

\item{condition.b}{Condition(s) in the group column that you want to test against.}

\item{fitType}{Type of fit to use with DESeq2 - parametric, local (Default) and mean.}

\item{ngenes}{Number of genes you want to test (Default: all genes)}
}
\description{
Wrapper for DESeq2.
}
\details{
This is a wrapper for DESeq2. The function handles the conversion of the 
\linkS4class{EMSet} to a DESeqDataset object for use with the DESeq2 
functions.
}
