% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_diffexpression.R
\name{RunDiffExpression}
\alias{RunDiffExpression}
\title{RunDiffExpression}
\usage{
RunDiffExpression(x, condition.a = NULL, condition.b = "Other",
  condition.list = NULL)
}
\arguments{
\item{x}{An expression matrix stored in either an \linkS4class{AEMSet} or a data.frame}

\item{condition.a}{A one-word string describing the main condition}

\item{condition.b}{A one-word string describing cells that do not fulfil condition A}

\item{condition.list}{A named list featuring individual cell identifiers and their associated condition.}
}
\description{
Runs differential expression on a pre-defined set of conditions. This function can only handle two conditions at a time, where Condition A describes a subset of cells and Condition B describes the remainder.
}
