% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_diffexpression.R
\name{RunPairedDE}
\alias{RunPairedDE}
\title{RunPairedDE}
\usage{
RunPairedDE(x, condition.a = NULL, condition.b = "Other",
  condition.list = NULL)
}
\description{
Called by the main function. Runs DESeq on one condition vs others at a time.
}
