% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_dimreduction.R
\name{RunPCA}
\alias{RunPCA}
\title{RunPCA}
\usage{
RunPCA(object, ngenes = 1500, scaling = TRUE)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object that has undergone filtering and normalisation.}

\item{ngenes}{The top number of genes you would like to perform the reduction by.}

\item{scaling}{Boolean - set to FALSE if you do not want to scale your values}
}
\description{
RunPCA
}
