% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_sctransform.R
\name{runSCTransform}
\alias{runSCTransform}
\title{runSCTransform}
\usage{
runSCTransform(object, batch = TRUE, seed = 1, verbose = FALSE,
  ngenes = 3000, ncells = NULL, ...)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}}}

\item{batch}{Normalise batches (Default: TRUE)}

\item{seed}{Set seed (Default: 1)}

\item{verbose}{Print messages from vst (Default: FALSE)}

\item{ngenes}{Number of genes to use for estimation of initial parameters 
(Default: 3000)}

\item{ncells}{Number of cells to use for the estimation of initial parameters
(Default: NULL)}

\item{...}{Parameters to pass on to the vst function}
}
\value{
An \code{\linkS4class{EMSet}} with normalised, corrected UMI counts 
in the normcounts and logcounts slots, and residuals in the residuals slot.
}
\description{
Normalise an \code{\linkS4class{EMSet}} with \pkg{sctransform}'s vst function
by Hafemeister & Satija 2019.
}
\details{
This function is a wrapper for the vst function from the sctransform
package. Users may supply additional functions to the vst function.
}
