% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_dimreduction.R
\name{RunTSNE}
\alias{RunTSNE}
\title{RunTSNE}
\usage{
RunTSNE(x, PCA = FALSE, dimensions = 2)
}
\arguments{
\item{x}{An expression matrix than can either be a full count matrix or a PCA-reduced matrix.}

\item{PCA}{Set this PCA flag to true if you are using a PCA-reduced matrix.}

\item{dimensions}{Number of dimensions you would like to reduce to}
}
\description{
RunTSNE
}
