% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_dimreduction.R
\docType{methods}
\name{runUMAP}
\alias{runUMAP}
\alias{runUMAP,EMSet-method}
\title{runUMAP}
\usage{
runUMAP(object, ..., method, config)

\S4method{runUMAP}{EMSet}(object, ..., method = c("naive", "umap-learn"),
  config = NULL)
}
\arguments{
\item{object}{An EMSet}

\item{...}{Additional arguments to pass to the UMAP function}

\item{method}{Method to use with the UMAP function - naive (default) or 
umap-learn (Python package required)}

\item{config}{Configuration to use with UMAP function (Optional)}
}
\value{
UMAP matrix stored in "UMAP" slot in reducedDims
}
\description{
Wrapper for the \code{\link[umap]{umap}} function.
}
\examples{
# Load example EMSet
em_set <- ascend::analyzed_set

# Run UMAP
em_set <- runUMAP(em_set, method = "naive")

}
