% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\docType{methods}
\name{SCE2EMSet,SingleCellExperiment-method}
\alias{SCE2EMSet,SingleCellExperiment-method}
\title{SCE2EMSet}
\usage{
\S4method{SCE2EMSet}{SingleCellExperiment}(x)
}
\arguments{
\item{x}{SingleCellExperiment object}
}
\value{
An EMSet
}
\description{
Converts an SingleCellExperiment object back into an EMSet. This function 
rebuilds the slots from data kept in the SingleCellExperiment. This should
not be used to initiate a new EMSet from a SingleCellExperiment.
}
