% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_scran.R
\name{SCESet2EMSet}
\alias{SCESet2EMSet}
\title{SCESet2EMset}
\usage{
SCESet2EMSet(SCESet, EMSet)
}
\arguments{
\item{SCESet}{A \linkS4class{SCESet} from \pkg{scater}}

\item{EMSet}{An \linkS4class{EMSet}}
}
\value{
An \linkS4class{EMSet} with data retrieved from a SCESet
}
\description{
Loads data from a SCESet to a pre-existing 
\linkS4class{EMSet} object.
}
