% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_scran.R
\name{SCESetnormalise}
\alias{SCESetnormalise}
\title{SCESetnormalise}
\usage{
SCESetnormalise(sce.set, em.set, quickCluster = FALSE)
}
\arguments{
\item{sce.set}{A \linkS4class{SingleCellExperiment} object}

\item{em.set}{An \linkS4class{EMSet} that the sce.set originated from}

\item{quickCluster}{Whether or not to use quickCluster}
}
\description{
Called by scranNormalise - runs normalisation on \pkg{SingleCellExperiment}
object and converts it back to EMSets.
}
