% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\docType{methods}
\name{show,AEMSet-method}
\alias{show,AEMSet-method}
\title{show}
\usage{
\S4method{show}{AEMSet}(object)
}
\description{
Displays a summary of the S4 object \linkS4class{AEMSet}
}
\details{
AEMSet summary
\itemize{
\item{\strong{Expression Matrix} - Number of cells and genes in the original expression matrix.}
\item{\strong{Controls} - Number of genes listed in each control group.}
\item{\strong{Filtering} - Details of any filtering the object has undergone.}
\item{\strong{Samples} - Number of samples/batches in the experiment.}
\item{\strong{Clusters} - If generated, the number of clusters and their associated sizes.}
\item{\strong{Log} - A record of methods used on the \linkS4class{AEMSet}.}
}
}
