% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\docType{methods}
\name{show,EMSet-method}
\alias{show,EMSet-method}
\title{show}
\usage{
\S4method{show}{EMSet}(object)
}
\description{
Generic display function for \linkS4class{EMSet}s. Displays summary of the
object.
}
