% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_jweb_charges.R
\name{split_jweb_charges}
\alias{split_jweb_charges}
\title{Split Justice Web charges into separate rows}
\usage{
split_jweb_charges(.data, .user_col, .charges_col)
}
\arguments{
\item{.data}{An R object containing the Justice Web data}

\item{.user_col}{Character string naming a column in \code{.data}
that contains the unique identifiers for each user}

\item{.charges_col}{Character string naming a column in \code{.data}
that contains the charges filed for an encounter}
}
\value{
A \code{data.frame}-class object containing one charge per row
}
\description{
Split Justice Web charges into separate rows
}
