% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getters.R
\docType{methods}
\name{[,EMSet,ANY,ANY,ANY-method}
\alias{[,EMSet,ANY,ANY,ANY-method}
\title{Subsetter for EMSet}
\usage{
\S4method{[}{EMSet,ANY,ANY,ANY}(x, i, j, drop = TRUE)
}
\arguments{
\item{x}{An \linkS4class{EMSet}}

\item{i}{Row index}

\item{j}{Col index}

\item{drop}{Drop stuff}
}
\description{
Subsetter for EMSet
}
