% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_setters.R
\docType{methods}
\name{[<-,EMSet,ANY,ANY,EMSet-method}
\alias{[<-,EMSet,ANY,ANY,EMSet-method}
\title{Replace entries of EMSet}
\usage{
\S4method{[}{EMSet,ANY,ANY,EMSet}(x, i, j, ...) <- value
}
\arguments{
\item{x}{EMSet}

\item{i}{Row index}

\item{j}{Column index}

\item{...}{...}

\item{value}{Replacement dataframe}
}
\description{
Replace entries of EMSet
}
