% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetmethods.R
\docType{methods}
\name{SubsetBatch,EMSet-method}
\alias{SubsetBatch,EMSet-method}
\title{SubsetBatch}
\usage{
\S4method{SubsetBatch}{EMSet}(object, batches = c())
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}}.}

\item{batches}{Name or number of the batch(es) you would like to subset.}
}
\value{
An \code{\linkS4class{EMSet}} containing only this batch.
}
\description{
Subset a specific batch from an \code{\linkS4class{EMSet}}.
}
\examples{
\dontrun{
sample1_emset <- SubsetBatch(em.set, batches = c("1", "2", "3"))
}
}
