% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetmethods.R
\name{SubsetBatch}
\alias{SubsetBatch}
\title{SubsetBatch}
\usage{
SubsetBatch(object, batches)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}}.}

\item{batches}{Name or number of the batch(es) you would like to subset.}
}
\description{
Generic for \code{\link{SubsetBatch}}.
}
