% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetmethods.R
\docType{methods}
\name{SubsetCells,EMSet-method}
\alias{SubsetCells,EMSet-method}
\title{SubsetCells}
\usage{
\S4method{SubsetCells}{EMSet}(object, cell.barcodes = c())
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}}}

\item{cell.barcodes}{A list of cell identifiers to subset from the 
\linkS4class{EMSet}}
}
\value{
An \code{\linkS4class{EMSet}}
}
\description{
Subset cells in the supplied list from an \code{\linkS4class{EMSet}}.
}
\examples{
\dontrun{
barcode_list <- c("Barcode1", "Barcode2", "Barcode3")
subset_emset <- SubsetCells(em.set, cell.barcodes = barcode_list)
}
}
