% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetmethods.R
\name{SubsetCells}
\alias{SubsetCells}
\title{SubsetCells}
\usage{
SubsetCells(object, cell.barcodes)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}}.}

\item{cell.barcodes}{A list of cell identifiers to subset from the}
}
\description{
Generic for \code{\link{SubsetCells}}.
}
