% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetmethods.R
\docType{methods}
\name{SubsetCluster,EMSet-method}
\alias{SubsetCluster,EMSet-method}
\title{SubsetCluster}
\usage{
\S4method{SubsetCluster}{EMSet}(object, clusters = c())
}
\arguments{
\item{object}{A \code{\linkS4class{EMSet}}.}

\item{clusters}{Clusters to subset from the dataset.}
}
\value{
Returns an \code{\linkS4class{EMSet}} containing only the selected 
clusters.
}
\description{
Subset an\code{\linkS4class{EMSet}} by cluster. Please make sure you have clustered 
with \code{\link{RunCORE}} before using this function.
}
\details{
This data is already normalised, but if you wish to recluster the data, you 
will need to use the RunPCA function again.
}
\examples{
\dontrun{
cluster1_object <- SubsetCluster(em.set, clusters = c("1", "2", "3"))
}
}
