% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_methods.R
\name{SubsetCluster}
\alias{SubsetCluster}
\title{SubsetCluster}
\usage{
SubsetCluster(object, x)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object}

\item{x}{Number of the cluster you would like to subset.}
}
\value{
Returns an \linkS4class{AEMSet} containing only this cluster.
}
\description{
Subset a \linkS4class{AEMSet} object by cluster. Please make sure you have clustered with \code{\link{FindOptimalClusters}} before using this function.
}
