% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetters.R
\docType{methods}
\name{subsetCondition,EMSet-method}
\alias{subsetCondition,EMSet-method}
\title{subsetCondition}
\usage{
\S4method{subsetCondition}{EMSet}(object, by = c(),
  conditions = list())
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}}}

\item{by}{A list containing parameter to subset object by. These should be 
present as a column in colInfo and names in the named list `conditions`.}

\item{conditions}{Conditions you would like to select. They should be 
organised into a named list, with the values specified in `by` as sublist
names.}
}
\value{
An \code{\linkS4class{EMSet}} containing cells that have these 
conditions.
}
\description{
Subset cells of specific conditions from an \code{\linkS4class{EMSet}}.
}
\examples{
# Load a pre-existing EMSet

}
