% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\docType{methods}
\name{SyncSlots,AEMSet-method}
\alias{SyncSlots,AEMSet-method}
\title{SyncSlots}
\usage{
\S4method{SyncSlots}{AEMSet}(object)
}
\description{
Synchronises the data frames ExpressionMatrix, GeneInformation and CellInformation, to ensure everything is up to date.
This is important as the object will undergo a series of changes during the filtering and normalisation process.
}
