% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{SyncSlots}
\alias{SyncSlots}
\title{SyncSlots (Generic)}
\usage{
SyncSlots(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet}}
}
\description{
Generic for the \code{\link{SyncSlots}} function.
}
