% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_utilities.R
\name{UnLog2Matrix}
\alias{UnLog2Matrix}
\title{UnLog2Matrix}
\usage{
UnLog2Matrix(expression.matrix)
}
\arguments{
\item{expression.matrix}{An expression.matrix}
}
\description{
Converts count values of a matrix that had undergone normalisation via scran's deconvolution method.
}
