% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\name{UpdateBatchInfo}
\alias{UpdateBatchInfo}
\title{UpdateBatchInfo}
\usage{
UpdateBatchInfo(object)
}
\description{
Can be called by the user or by a filtering function. Updates Batch Information in a \linkS4class{AEMSet} object, removing cells that have been filtered out.
}
