% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\name{UpdateControls}
\alias{UpdateControls}
\title{UpdateControls}
\usage{
UpdateControls(object, gene.list)
}
\arguments{
\item{object}{An \linkS4class{AEMSet} object.}

\item{gene.list}{A named list containing the gene identifiers. These identifiers must match the identifiers used in the expression matrix.}
}
\description{
Replaces the control list in a \linkS4class{AEMSet} object with a new control list. This also recalculates the metrics associated with the \linkS4class{AEMSet} object.
For best results, define your controls before you attempt any filtering.
}
