% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\docType{methods}
\name{updateObject,EMSet-method}
\alias{updateObject,EMSet-method}
\title{updateObject}
\usage{
\S4method{updateObject}{EMSet}(object)
}
\arguments{
\item{object}{An old EMSet.}
}
\value{
An updated EMSet.
}
\description{
Converts EMSets from ascend version < 0.5.0 to new EMSet that inherits from
SingleCellExperiment. Please note that normalised data will be used as the 
main count matrix, and loaded into the normcounts slot if a normalisation
method was logged. Quality control metrics will be re-calculated upon 
conversion.
}
\examples{
\dontrun{
# Make sure you replace your original object with the updated object
object <- updateObject(object)
}
}
