% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetters.R
\docType{methods}
\name{useVariableGenes,EMSet-method}
\alias{useVariableGenes,EMSet-method}
\title{useVariableGenes}
\usage{
\S4method{useVariableGenes}{EMSet}(object, ngenes = 1500)
}
\description{
Reduces an EMSet to a range of the most variable genes as determined by 
the values calculated by \code{\link{calculateCV}}.
}
