% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_diffexpression.R
\name{VerifyArguments}
\alias{VerifyArguments}
\title{VerifyArguments}
\usage{
VerifyArguments(condition.a = NULL, condition.b = NULL,
  condition.list = list())
}
\description{
Called by RunDiffExpression. Makes sure conditions are okay.
}
