% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitepaper.R
\name{whitepaper}
\alias{whitepaper}
\title{Ascend Innovations Whitepaper Template}
\source{
Literally copied the format of the Ascend Innovations whitepapers
}
\usage{
whitepaper(
  ...,
  keep_tex = FALSE,
  md_extensions = c("-autolink_bare_uris"),
  fig_caption = TRUE,
  toc = TRUE,
  template = "wp_template.tex",
  format = "whitepaper",
  highlight = "default",
  xtab_caption_placement = "top",
  xtab_comment = FALSE
)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Should the raw TeX files be kept?}

\item{md_extensions}{Extensions passed to the Markdown engine}

\item{fig_caption}{Should figure captions be enabled?}

\item{toc}{Should a table of contents be created?}

\item{template}{The TeX template used}

\item{format}{The format of the document}

\item{highlight}{Highlight color template for code chunks}

\item{xtab_caption_placement}{Placement of captions for tables created using the \code{xtable} package}

\item{xtab_comment}{Logical variable whether the \code{xtable} timestamp should be printed}
}
\value{
R Markdown output format to pass to \code{rmarkdown::render}
}
\description{
Create an Ascend Innovations whitepaper using 
             \code{rmarkdown} and \code{bookdown}
}
\examples{

\dontrun{
rmarkdown::draft("my_whitepaper.Rmd", template = "whitepaper", package = "Ascend")
}

}
