% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{z_theme}
\alias{z_theme}
\title{z_theme}
\usage{
z_theme()
}
\value{
A ggplot2 theme to use with ggplots
}
\description{
From perceptions - https://github.com/zonination/perceptions. This function 
sets the colour themes for use with the \code{\link{PlotTopGeneExpression}} 
function.
}
