% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\docType{methods}
\name{addControlInfo,EMSet-method}
\alias{addControlInfo,EMSet-method}
\title{addControlInfo}
\usage{
\S4method{addControlInfo}{EMSet}(x, controls = NULL)
}
\arguments{
\item{x}{An \linkS4class{EMSet}.}

\item{controls}{A named list of control groups and controls.}
}
\value{
An \linkS4class{EMSet} with controls integrated into \code{rowInfo}
and calculated QC metrics.
}
\description{
Adds and integrates control information to an \linkS4class{EMSet}.
}
\examples{
# Load data
load("R/sysdata.rda")
x <- data_package$raw_emset

# Create a new list of controls
new_control <- list(control3 = c("Gene3", "Gene4"))

# Add to controls
updated_x <- addControlInfo(x, controls = new_control)

}
