% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\docType{class}
\name{AEMSet-class}
\alias{AEMSet-class}
\title{ASCEND Expression and Metadata Set (AEMSet)}
\description{
An S4 class to contain data in a format ASCEND can work with for analysis.
}
\section{Slots}{

\describe{
\item{\code{ExpressionMatrix}}{Transcript counts stored as a sparse matrix, where rows are transcript/gene identifiers and columns are invididual cells.}

\item{\code{GeneAnnotation}}{A data frame containing information on genes and their corresponding identifiers, such as gene names and ENSEMBL transcript identifiers.}

\item{\code{BatchInformation}}{A named list containing each cell identifier and its associated batch/sample.}

\item{\code{Controls}}{A named list featuring gene identifiers to use as controls. These gene identifiers must match the identifiers used in the expression matrix.}

\item{\code{Conditions}}{A named list containing a list of conditions and what category each cell identifier falls into.}

\item{\code{PCA}}{Objects related to dimension reduction, such as a PCA matrix, a reduced PCA matrix and a list of percentage variance values per principle component. Populated by \code{\link{RunPCA}}.}

\item{\code{TSNE}}{TSNE-reduced matrices used for plotting. Populated by \code{\link{PlotTSNE}}.}

\item{\code{Clusters}}{Objects related to clustering, including a distance matrix, a hclust object, cell identifiers and their associated cluster. Populated by \code{\link{FindOptimalClusters}}.}

\item{\code{DifferentialExpression}}{Differential expression results from DESeq. Populated by \code{\link{RunDiffExpression}} and \code{\link{RunClusterDiffExpression}}.}

\item{\code{Metrics}}{A list of values generated by the \code{\link{GenerateMetrics}} function.}

\item{\code{Log}}{A record of functions used on an \linkS4class{AEMSet}.}
}}

