% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_clustering.R
\name{BuildKeyStat}
\alias{BuildKeyStat}
\title{BuildKeyStat
Generates a bunch of values based on other calculates. Called by 
\code{\link{RunCORE}}.}
\usage{
BuildKeyStat(rand.matrix = NULL, nres = 40)
}
\arguments{
\item{rand.matrix}{Rand matrix generated by RunCORE.}

\item{nres}{Number of resolutions inputted by user.}
}
\value{
A labelled dataframe.
}
\description{
BuildKeyStat
Generates a bunch of values based on other calculates. Called by 
\code{\link{RunCORE}}.
}
