% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_cellranger.R
\name{CellRangerToASCEND}
\alias{CellRangerToASCEND}
\title{CellRangerToASCEND}
\usage{
CellRangerToASCEND(cellranger.path, genome.name)
}
\arguments{
\item{cellranger.path}{This is the path to the folder of data generated by Cell Ranger's count pipeline.}

\item{genome}{This is the name of the reference genome used by Cell Ranger.}
}
\value{
\linkS4class{AEMSet}
}
\description{
\code{\link{CellRangerToASCEND}} processes data generated by Cell Ranger's count pipeline into a \linkS4class{AEMSet} object, for use with \pkg{ASCEND}.
This function assumes the user is using ribosomal and mitochondrial genes as a control for the experiment.
}
