% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{CheckControls}
\alias{CheckControls}
\title{CheckControls
Another check function for EMSets. This check is only run if controls are 
present.}
\usage{
CheckControls(object, errors)
}
\arguments{
\item{object}{An EMSEt object.}

\item{errors}{Vector to store error messages in. If the object is returned
with values - the object will be declared invalid.}
}
\description{
CheckControls
Another check function for EMSets. This check is only run if controls are 
present.
}
