% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_setters.R, R/ascend_getters.R
\docType{methods}
\name{colInfo<-}
\alias{colInfo<-}
\alias{colInfo<-,EMSet-method}
\alias{colInfo}
\alias{colInfo,EMSet-method}
\title{colInfo}
\usage{
colInfo(x, ...) <- value

\S4method{colInfo}{EMSet}(x) <- value

colInfo(x, ...)

\S4method{colInfo}{EMSet}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{\linkS4class{EMSet}}

\item{...}{...}

\item{value}{DataFrame to store in colInfo slot.}

\item{withDimnames}{Retrieve rownames of DataFrame Default: TRUE}
}
\value{
colInfo DataFrame
}
\description{
Access cell information stored in colInfo.
}
