% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getters.R, R/ascend_setters.R
\docType{methods}
\name{controls}
\alias{controls}
\alias{controls,EMSet-method}
\alias{controls<-}
\alias{controls<-,EMSet-method}
\title{controls}
\usage{
controls(x, ...)

\S4method{controls}{EMSet}(x)

controls(x, ...) <- value

\S4method{controls}{EMSet}(x) <- value
}
\arguments{
\item{x}{\linkS4class{EMSet}}

\item{...}{...}

\item{value}{Named list of control genes}
}
\value{
EMSet with defined controls and qc metrics associated with controls
}
\description{
Access set controls.
}
\examples{
# Load EMSet
em_set <- ascend::raw_set

# Access controls
controls <- controls(em_set)

# Set controls
controls(em_set) <- controls

}
