% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\name{ConvertGeneAnnotation}
\alias{ConvertGeneAnnotation}
\title{ConvertGeneAnnotation}
\usage{
ConvertGeneAnnotation(object, old.annotation, new.annotation)
}
\description{
Generic method for \code{\link{ConvertGeneAnnotation}}.
}
