% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_methods.R
\name{ConvertGeneAnnotation}
\alias{ConvertGeneAnnotation}
\title{ConvertGeneAnnotation}
\usage{
ConvertGeneAnnotation(object, old.annotation, new.annotation)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object.}

\item{old.annotation}{Name of the column containing the current gene annotations.}

\item{new.annotation}{Name of the column you would like to convert the gene annotations to.}
}
\description{
Convert gene identifiers used in this \linkS4class{AEMSet} to identifiers used in another column of a dataframe stored in the GeneAnnotation slot.
}
