% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\docType{methods}
\name{convertGeneID}
\alias{convertGeneID}
\alias{convertGeneID,EMSet-method}
\title{convertGeneID}
\usage{
convertGeneID(object, ..., new.annotation)

\S4method{convertGeneID}{EMSet}(object, ..., new.annotation = NULL)
}
\arguments{
\item{object}{\linkS4class{EMSet}}

\item{...}{...}

\item{new.annotation}{Name of the column where the new set of identifiers 
are stored in rowInfo}
}
\description{
Switches gene identifiers used in the EMSet to the identifiers stored in a 
column of rowInfo.
}
\examples{
# Load data
x <- ascend::raw_set

# Extract rowInfo from EMSet
row_info <- rowInfo(x)

# Create new set of identifiers
ensembl_id <- paste0("ENSG0000000000", 1:nrow(row_info))
row_info$ensembl_id <- ensembl_id

# Replace rowInfo in EMSet
rowInfo(x) <- row_info

# Switch identifiers
switched_annotations <- convertGeneID(x, new.annotation = "ensembl_id")

}
