% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_scran.R
\name{ConvertScaterToascend}
\alias{ConvertScaterToascend}
\title{ConvertScaterToascend}
\usage{
ConvertScaterToascend(SCESet, EMSet)
}
\arguments{
\item{SCESet}{A \linkS4class{SCESet} from \pkg{scater}}

\item{EMSet}{An \linkS4class{EMSet}}
}
\description{
Loads data from a SCATER object to a pre-existing ascend object.
}
