% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_scran.R
\name{ConvertToSCE}
\alias{ConvertToSCE}
\title{ConvertToSCE}
\usage{
ConvertToSCE(object, control.list = list())
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object.}

\item{control.list}{Optional - a named list containing mitochondrial and 
ribosomal genes.}
}
\value{
A SingleCellExperiment object
}
\description{
Convert a \code{\linkS4class{EMSet}} object into a SingleCellExperiment object for 
use with \pkg{scater}, \pkg{scran} and other Bioconductor packages.
}
\details{
In order to use this function, you must have mitochondrial and ribosomal 
genes in your expression data.
}
\examples{
\dontrun{
single_cell_experiment <- ConvertToSCE(em.set, control.list = list(
Mt = mt.genes, Rb = rb.genes
))}
}
