% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\docType{methods}
\name{DisplayLog,EMSet-method}
\alias{DisplayLog,EMSet-method}
\title{DisplayLog}
\usage{
\S4method{DisplayLog}{EMSet}(x)
}
\arguments{
\item{x}{An \linkS4class{EMSet}}
}
\value{
A print-out of the log attached to an \code{\linkS4class{EMSet}}.
}
\description{
Print out the log of an \code{\linkS4class{EMSet}}.
}
