% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\name{DisplayLog}
\alias{DisplayLog}
\title{DisplayLog}
\usage{
DisplayLog(object)
}
\value{
A print-out of the log attached to an \linkS4class{EMSet}
}
\description{
Print out the log of an \linkS4class{EMSet}.
}
