% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\docType{class}
\name{EMSet-class}
\alias{EMSet-class}
\alias{.EMSet}
\title{Expression and Metadata Set (EMSet)}
\description{
An \linkS4class{EMSet} is a S4 class that stores data in a format `ascend`` can 
easily work with for analysis.
}
\details{
This iteration of the \linkS4class{EMSet} inherits from the 
\code{\linkS4class{SingleCellExperiment}} superclass for integration into
Bioconductor.

`ascend`-specific slots are as follows:
}
\section{Slots}{

\describe{
\item{\code{colInfo}}{A data frame containing each cell identifier, its 
associated batch/sample and additional information such as conditions. This 
slot was called CellInformation in earlier versions of `ascend`.}

\item{\code{rowInfo}}{A data frame containing information a set of gene 
identifiers, such as gene symbols or ENSEMBL transcript identifiers. This 
data frame also holds information on controls and any information provided by 
the user. This slot was called GeneInformation in earlier versions of `ascend`.}

\item{\code{clusterAnalysis}}{Objects related to clustering - distance matrix and
hclust object}

\item{\code{log}}{A record of functions used on an \linkS4class{EMSet}.

The following data is loaded into \linkS4class{SingleCellExperiment} slots:}

\item{\code{assays}}{Raw expression matrix is loaded into the \code{counts} slot.
Normalised counts are loaded into the following based on their normalisation
method:
 \describe{
  \item{normcounts}{Counts normalised by `ascend`'s RLE method.}
  \item{logcounts}{Counts normalised by `scran`'s deconvolution method.}
}}

\item{\code{colData}}{The following cell-based QC metrics are stored in this 
DataFrame:
\describe{
\item{qc_libsize}{Total number of transcripts per cell.}
\item{qc_ngenes}{Total number of expressed genes per cell.}
\item{qc_control_expression}{Percentage of control expression to total 
expression.}
}}

\item{\code{rowData}}{The following gene-based QC metrics are stored in this
DataFrame:
\describe{
\item{qc_ncounts}{Total gene expression across dataset.}
\item{qc_ncells}{Number of cells the gene is expressed in.}
\item{qc_meancounts}{Mean expression level of the gene.}
\item{qc_topgeneranking}{Position of gene in top gene expression rankings.}
\item{qc_pct_total_expression}{Proportion of gene expression to total
expression.}
}}
}}

