% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\docType{methods}
\name{ExcludeControl,EMSet-method}
\alias{ExcludeControl,EMSet-method}
\title{ExcludeControl}
\usage{
\S4method{ExcludeControl}{EMSet}(x, control.name)
}
\arguments{
\item{x}{An \code{\linkS4class{EMSet}} object. It is recommended that 
you run this step after this object has undergone filtering.}

\item{control.name}{Name of the control set you want to remove from the 
dataset.}
}
\value{
\code{x} without the \code{control.name} set of controls in the 
expression matrix.
}
\description{
Removes the specified control from the expression matrix. This should be used
after normalisation, and prior to PCA reduction, clustering and differential
expression analysis.
}
