% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_methods.R
\name{ExcludeControl}
\alias{ExcludeControl}
\title{ExcludeControl}
\usage{
ExcludeControl(x, control.name)
}
\arguments{
\item{x}{An \code{\linkS4class{EMSet}} object. It is recommended that 
you run this step after this object has undergone filtering.}

\item{control.name}{Name of the control set you want to remove from the dataset.}
}
\description{
Generic method for \code{\link{ExcludeControl}}.
}
