% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetters.R
\docType{methods}
\name{excludeControl}
\alias{excludeControl}
\alias{excludeControl,EMSet-method}
\title{excludeControl}
\usage{
excludeControl(object, ..., control)

\S4method{excludeControl}{EMSet}(object, control = c())
}
\arguments{
\item{object}{An \linkS4class{EMSet}.}

\item{...}{...}

\item{control}{Name of the control group you would like to exclude. It must
be in the `control` column of rowInfo.}
}
\value{
An \linkS4class{EMSet} without the specified control group.
}
\description{
Excludes a control group from an \linkS4class{EMSet}.
}
\examples{
# Load example EMSet
em_set <- ascend::raw_set

# Exclude Mitochondrial-related genes
em_set <- excludeControl(em_set, control = "Mt")

}
