% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_methods.R
\name{ExcludeControl}
\alias{ExcludeControl}
\title{ExcludeControl}
\usage{
ExcludeControl(object, control.name)
}
\arguments{
\item{object}{A \linkS4class{EMSet} object. It is recommended that you run this step after this object has undergone filtering.}

\item{control.name}{Name of the control set you want to remove from the dataset.}
}
\description{
Removes the specified control from the expression matrix.
}
