% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_utilities.R
\name{FileCheck}
\alias{FileCheck}
\title{FileCheck 
This little snippet is designed in a way in which you pair it up with a call 
to process the file. This checks if the file exists.}
\usage{
FileCheck(filename)
}
\arguments{
\item{filename}{Name of the file you want to check.}
}
\value{
Returns a boolean
}
\description{
FileCheck 
This little snippet is designed in a way in which you pair it up with a call 
to process the file. This checks if the file exists.
}
