% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_filtering.R
\name{filterByControl}
\alias{filterByControl}
\title{filterByControl}
\usage{
filterByControl(object, control = NULL, pct.threshold = 20)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}}}

\item{control}{Name of the control group, as used in the named list 
supplied to the EMSet object.}

\item{pct.threshold}{Percentage threshold to filter cells by, as a whole 
number. Default: 20.}
}
\value{
An \code{\linkS4class{EMSet}} with filtered controls.
}
\description{
Filter cells in an expression matrix based on the expression levels of a
specific control.
This function should be used AFTER the cells have undergone general filtering
with the \code{\link{filterByOutliers}} function.
}
\examples{
# Load EMSet
load("R/sysdata.rda")
raw_emset <- data_package$raw_emset

# Filter by outliers with default settings
filtered_emset <- filterByControl(raw_emset, control = "control1", pct.threshold = 20)

}
