% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_filtering.R
\name{FilterByControl}
\alias{FilterByControl}
\title{FilterByControl}
\usage{
FilterByControl(control.name = NULL, pct.threshold = 20, object)
}
\arguments{
\item{control.name}{Name of the control group, as used in the named list supplied to the AEMSet object}

\item{pct.threshold}{Percentage threshold to filter cells by, as a whole number. Default: 20}

\item{object}{A \linkS4class{AEMSet} object.}
}
\description{
Filter cells in an expression matrix based on the expression levels of a
specific control.
This function should be used AFTER the cells have undergone general filtering
with the \code{\link{FilterByOutliers}} function.
}
