% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_filtering.R
\name{FilterByControl}
\alias{FilterByControl}
\title{FilterByControl}
\usage{
FilterByControl(control.group, total.counts, expression.matrix)
}
\description{
Called by \code{\link{FilterByOutliers}}. This function identifies cells to remove based on expression levels of control genes.
}
