% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_filtering.R
\name{FilterByExpressedGenesPerCell}
\alias{FilterByExpressedGenesPerCell}
\title{FilterByExpressedGenesPerCell}
\usage{
FilterByExpressedGenesPerCell(object, pct.value = 1)
}
\arguments{
\item{object}{EMSet (ascend) object that has been filtered by
\code{\link{FilterByOutliers}} and \code{\link{FilterByControl}}.}

\item{pct.value}{Percentage threshold as a whole number. Default: 1}
}
\description{
Filtered cells by the proportion of expressed genes in the cell. This step is
usually done after the other filtering steps and prior to normalisation.
}
