% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_filtering.R
\name{FilterControl}
\alias{FilterControl}
\title{FilterControl}
\usage{
FilterControl(control.group, total.counts, expression.matrix)
}
\description{
Called by \code{\link{FilterByOutliers}}. This function identifies cells to
remove based on expression levels of control genes.
}
