% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_filtering.R
\name{FilterControl}
\alias{FilterControl}
\title{FilterControl}
\usage{
FilterControl(control.group, total.counts, expression.matrix)
}
\arguments{
\item{control.group}{Name of control group}

\item{total.counts}{List of total counts of each cell}

\item{expression.matrix}{An expression matrix}
}
\value{
Percentage expression of controls per cell
}
\description{
Called by \code{\link{FilterByOutliers}}. This function identifies cells to
remove based on expression levels of control genes.
}
