% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_filtering.R
\name{filterLowAbundanceGenes}
\alias{filterLowAbundanceGenes}
\title{filterLowAbundanceGenes}
\usage{
filterLowAbundanceGenes(object, pct.threshold = 1)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has been filtered by 
\code{\link{FilterByOutliers}} and \code{\link{FilterByControl}}.}

\item{pct.threshold}{Percentage threshold as a whole number. Default: 1.}
}
\value{
An \code{\linkS4class{EMSet}} with low abundance genes removed from 
the dataset.
}
\description{
Removes genes if they are expressed in less than a set percentage of cells . 
This step is usually done after the other filtering steps and prior to 
normalisation. As this step may remove rare transcripts, this filtering step
is optional.
}
\examples{

}
