% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_filtering.R
\name{filterLowAbundanceGenes}
\alias{filterLowAbundanceGenes}
\title{filterLowAbundanceGenes}
\usage{
filterLowAbundanceGenes(object, pct.threshold = 1)
}
\arguments{
\item{object}{An \linkS4class{EMSet} that has been filtered by 
\code{\link{filterByOutliers}} and \code{\link{filterByControl}}.}

\item{pct.threshold}{Percentage threshold as a whole number. Default: 1.}
}
\value{
An \linkS4class{EMSet} with low abundance genes removed from the 
dataset.
}
\description{
Removes genes if they are expressed in less than a set percentage of cells . 
This step is usually done after the other filtering steps and prior to 
normalisation. As this step may remove rare transcripts, this filtering step
is optional.
}
\examples{
# Load EMSet
raw_emset <- ascend::data_package$raw_emset

# Filter low abundance genes expressed in less than 1\% of cells
filtered_emset <- filterLowAbundanceGenes(raw_emset, pct.threshold = 1)

}
