% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_clustering.R
\name{FindOptimalResult}
\alias{FindOptimalResult}
\title{FindOptimalResult}
\usage{
FindOptimalResult(key.stats)
}
\description{
Finds the optimal result from trends in stability. Called by RunCORE.
}
