% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_filtering.R
\name{FindOutliers}
\alias{FindOutliers}
\title{FindOutliers}
\usage{
FindOutliers(values, nmads = 3, type = c("both", "lower", "upper"),
  na.rm = FALSE)
}
\arguments{
\item{values}{List of values}

\item{nmads}{Mean Absolute Deviation value threshold}

\item{type}{Direction to find outliers in - both, lower, upper}
}
\description{
Adapted from \code{\link[scater]{isOutlier}}. Determines outliers based on MAD value.
}
