% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_filtering.R
\name{findOutliers}
\alias{findOutliers}
\title{findOutliers}
\usage{
findOutliers(values, nmads = 3, type = c("both", "lower", "upper"),
  na.rm = FALSE)
}
\arguments{
\item{values}{List of values.}

\item{nmads}{Mean Absolute Deviation value threshold. Default: 3.}

\item{type}{Direction to find outliers in - both (Default), lower, upper.}

\item{na.rm}{Remove NA values if present (Default: False).}
}
\value{
A boolean of values that fall between the ranges set by \code{nmads} 
and \code{type}.
}
\description{
Adapted from \code{\link[scater]{isOutlier}}. Determines outliers based on 
Mean Absolute Deviation (MAD) value.
}
