% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_clustering.R
\name{GenerateClusteringMatrix}
\alias{GenerateClusteringMatrix}
\title{GenerateClusteringMatrix}
\usage{
GenerateClusteringMatrix(original.clusters, cluster.list)
}
\description{
Performs a series of cuts over 40 heights using dynamicTreeCut with a set height..
}
