% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_clustering.R
\name{GenerateRandIndexMatrix}
\alias{GenerateRandIndexMatrix}
\title{GenerateRandIndexMatrix}
\usage{
GenerateRandIndexMatrix(cluster.matrix, original.clusters, cluster.list)
}
\arguments{
\item{cluster.matrix}{A matrix containing clustering results generated by 
cutting the hclust object at 40 resolutions.}

\item{original.clusters}{Clusters generated by dynamicTreeCut.}

\item{cluster.list}{A list of clustering results to unpack.}
}
\description{
Generates columns in KeyStats data frame related to rand index calculations.
}
