% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_clustering.R
\name{GenerateStabilityValues}
\alias{GenerateStabilityValues}
\title{GenerateStabilityValues}
\usage{
GenerateStabilityValues(rand.idx.matrix = NULL, nres = 40)
}
\arguments{
\item{rand.idx.matrix}{Matrix generated by RunCORE}

\item{nres}{Number of resolutions}
}
\description{
Generates a series of stability values based on rand matrix result
}
