% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\docType{methods}
\name{GetBatchMatrix,EMSet-method}
\alias{GetBatchMatrix,EMSet-method}
\title{GetBatchMatrix}
\usage{
\S4method{GetBatchMatrix}{EMSet}(object, batch.id)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object.}

\item{batch.id}{Batch identifier that you would like to retrieve.}
}
\description{
Retrieve a portion of the matrix by batch label
}
