% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\name{GetBatchMatrix}
\alias{GetBatchMatrix}
\title{GetBatchMatrix}
\usage{
GetBatchMatrix(object, batch.id)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}

\item{batch.id}{Batch identifier that you would like to retrieve}
}
\description{
Generic for \code{\link{GetBatchMatrix}}.
}
