% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_clustering.R
\name{GetConsecutiveSequence}
\alias{GetConsecutiveSequence}
\title{GetConsecutiveSequence}
\usage{
GetConsecutiveSequence(x, direction = c("forward", "reverse"))
}
\description{
Function used to determine consecutive sequences on boundaries.
This is called by FindOptimalResult to determine which clustering results are
consecutive.
}
