% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\docType{methods}
\name{GetControls,EMSet-method}
\alias{GetControls,EMSet-method}
\title{GetControls}
\usage{
\S4method{GetControls}{EMSet}(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object.}
}
\value{
This function returns a list of controls defined by the user.
}
\description{
Retrieve list of controls from \code{\linkS4class{EMSet}}.
}
