% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\name{GetControls}
\alias{GetControls}
\title{GetControls}
\usage{
GetControls(object)
}
\arguments{
\item{object}{An \linkS4class{AEMSet} object.}
}
\description{
Retrieve list of controls from \linkS4class{AEMSet}.
}
