% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\name{GetControls}
\alias{GetControls}
\title{GetControls}
\usage{
GetControls(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}
}
\value{
This function returns a list of controls defined by the user.s
}
\description{
Generic for \code{\link{GetControls}}.
}
