% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\docType{methods}
\name{GetDistanceMatrix,EMSet-method}
\alias{GetDistanceMatrix,EMSet-method}
\title{GetDistanceMatrix}
\usage{
\S4method{GetDistanceMatrix}{EMSet}(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}
}
\value{
Distance matrix stored in \code{object}
}
\description{
Retrieve the distance matrix from an \code{\linkS4class{EMSet}} object that 
has undergone clustering.
}
\examples{
# Load EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", 
"ExampleClusteredEMSet.rds"))

# Retrieve distance matrix from the object
distanceMatrix <- GetDistanceMatrix(EMSet)

}
