% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\name{GetDistanceMatrix}
\alias{GetDistanceMatrix}
\title{GetDistanceMatrix}
\usage{
GetDistanceMatrix(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}
}
\value{
Distance matrix stored in \code{object}
}
\description{
Generic for \code{\link{GetDistanceMatrix}}.
}
