% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\name{GetDistanceMatrix}
\alias{GetDistanceMatrix}
\title{GetDistanceMatrix}
\usage{
GetDistanceMatrix(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object.}
}
\value{
This function returns a distance matrix.
}
\description{
Retrieve the distance matrix from an \linkS4class{EMSet} object that has undergone clustering.
}
