% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\name{GetExpressionMatrix}
\alias{GetExpressionMatrix}
\title{GetExpressionMatrix}
\usage{
GetExpressionMatrix(object, format)
}
\arguments{
\item{object}{An \linkS4class{EMSet}}

\item{format}{Format of the returned matrix ("data.frame", "matrix")}
}
\value{
Returns the expression matrix in the chosen format (data.frame or 
matrix)
}
\description{
Returns a data frame containing the expression matrix from an 
\linkS4class{EMSet}.
}
\examples{
# Load EMSet
EMSet <- readRDS(system.file(package = "ascend", 
"extdata", "ExampleEMSet.rds"))

# Retrieve expression matrix as a data frame
data_frame <- GetExpressionMatrix(EMSet, "data.frame")

# Retrieve expression matrix as a matrix
matrix <- GetExpressionMatrix(EMSet, "matrix")

}
