% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_getfunctions.R
\name{GetExpressionMatrix}
\alias{GetExpressionMatrix}
\title{GetExpressionMatrix}
\usage{
GetExpressionMatrix(object, format)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} to retrieve the expression matrix from}

\item{format}{Format of the returned matrix - 'data.frame' or 'matrix'}
}
\value{
Returns the expression matrix in the chosen format (data.frame or matrix).
}
\description{
Returns a data frame containing the expression matrix from a \linkS4class{AEMSet} object.
}
