% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\name{GetGeneAnnotation}
\alias{GetGeneAnnotation}
\title{GetGeneAnnotation}
\usage{
GetGeneAnnotation(object)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object.}
}
\description{
Retrieve gene annotation information from an \linkS4class{AEMSet} object.
}
