% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_getfunctions.R
\name{GetGeneInfo}
\alias{GetGeneInfo}
\title{GetGeneInfo}
\usage{
GetGeneInfo(object)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object.}
}
\value{
This function returns a data frame.
}
\description{
Retrieve gene information from an \linkS4class{AEMSet} object.
}
