% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\name{GetGeneInfo}
\alias{GetGeneInfo}
\title{GetGeneInfo}
\usage{
GetGeneInfo(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}
}
\value{
This function returns a data frame containing information about 
the genes in \code{object}.
}
\description{
Generic for \code{\link{GetGeneInfo}}.
}
\examples{
# Load EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", "ExampleEMSet.rds"))

# Retrieve data frame containing gene information from the object
gene.info <- GetGeneInfo(EMSet)

}
