% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_getfunctions.R
\name{GetHclust}
\alias{GetHclust}
\title{GetHclust}
\usage{
GetHclust(object)
}
\arguments{
\item{object}{An \linkS4class{AEMSet} object.}
}
\value{
This function returns a hclust object.
}
\description{
Retrieve the Hclust object from an \linkS4class{AEMSet} object that has undergone clustering.
}
