% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\name{GetHclust}
\alias{GetHclust}
\title{GetHclust}
\usage{
GetHclust(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object}
}
\description{
Generic for \code{\link{GetHclust}}.
}
