% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_getfunctions.R
\name{GetPCA}
\alias{GetPCA}
\title{GetPCA}
\usage{
GetPCA(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object.}
}
\value{
This function returns a data frame.
}
\description{
Retrieve the PCA matrix from an \linkS4class{EMSet} object that has undergone PCA reduction.
}
