% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_getfunctions.R
\name{GetRandMatrix}
\alias{GetRandMatrix}
\title{GetRandMatrix}
\usage{
GetRandMatrix(object)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object that has undergone clustering.}
}
\value{
This function returns a data frame.
}
\description{
Retrieve the rand matrix used to determine clusters from an \linkS4class{AEMSet} object.
}
