% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getfunctions.R
\name{GetRandMatrix}
\alias{GetRandMatrix}
\title{GetRandMatrix}
\usage{
GetRandMatrix(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object that has undergone 
clustering}
}
\description{
Generic for \code{\link{GetRandMatrix}}.
}
