% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_cellranger.R
\name{LoadCellRanger}
\alias{LoadCellRanger}
\title{LoadCellRanger}
\usage{
LoadCellRanger(cellranger.path, genome.name)
}
\arguments{
\item{cellranger.path}{This is the path to the folder of data generated by
Cell Ranger's count pipeline}

\item{genome}{This is the name of the reference genome used by Cell Ranger}
}
\value{
An \linkS4class{EMSet} with cell information, gene information and
controls set
}
\description{
\code{\link{LoadCellRanger}} processes data generated by Cell Ranger's
count pipeline into a \linkS4class{EMSet} object, for use with \pkg{ascend}.
This function assumes the user is using ribosomal and mitochondrial genes as
a control for the experiment.
}
\examples{
\dontrun{
new_em_set <- LoadCellRanger("CellRangerData/", "GRCh38")
}
}
