% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_cellranger.R
\name{LoadCellRangerData}
\alias{LoadCellRangerData}
\title{LoadCellRangerData}
\usage{
LoadCellRangerData(working.dir, genome.name)
}
\arguments{
\item{working.dir}{Path inputted by user.}

\item{genome.name}{Name of reference inputted by the user.}
}
\description{
Called by \code{\link{LoadCellRanger}}. This function detects which folder
to use in the 'outs' directory of the Cell Ranger output folder.
}
