% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{newEMSet}
\alias{newEMSet}
\title{newEMSet}
\usage{
newEMSet(x, colInfo = NULL, colData = NULL, rowInfo = NULL,
  rowData = NULL, controls = NULL)
}
\arguments{
\item{x}{An object containing count data. The counts can be stored in the 
following formats: SingleCellExperiment, "counts" in a list or SimpleList,
dense matrix or sparse matrix.}

\item{colInfo}{A data frame containing cell-related metadata.}

\item{colData}{A data frame containing cell-related data.}

\item{rowInfo}{A data frame containing gene-related metadata.}

\item{rowData}{A data frame containing gene-related data.}

\item{controls}{A named list containing control genes grouped into named 
control groups.}
}
\value{
An \linkS4class{EMSet}.
}
\description{
\code{\link{newEMSet}} generates a \linkS4class{EMSet} object for use with 
the `ascend` package. This object contains an expression matrix, 
associated metadata, downstream analysis and a log documenting the actions 
used to shape the data in this object.
}
\examples{
# Randomly generate count matrix
count_matrix <- matrix(sample(0:1, 100, replace=TRUE),10,10)

# Generate cell barcodes
cell_barcodes <- paste0("Cell-", 1:10)
gene_ids <- paste0("Gene-", 1:10)

# Add to matrix
colnames(count_matrix) <- cell_barcodes
rownames(count_matrix) <- gene_ids

# Create an EMSet
# EMSet from a list
em_set <- newEMSet(list(counts = count_matrix))

# EMSet from a matrix
em_set <- newEMSet(count_matrix)


}
