% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{normaliseBatches}
\alias{normaliseBatches}
\title{normaliseBatches}
\usage{
normaliseBatches(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} with cells from more than one batch.}
}
\value{
An \code{\linkS4class{EMSet}} with batch-normalised expression values.
}
\description{
Normalise counts to remove batch effects. This normalisation method is for
experiments where data from batches of different samples are combined without
undergoing library equalisation.
}
\details{
This step should be done prior to any filtering and normalisation between cells.
}
\examples{
# Generate example matrix
count_matrix <- matrix(sample(0:1, 900, replace=TRUE),30,30)
colnames(count_matrix) <- paste0("Cell-", 1:ncol(count_matrix))
rownames(count_matrix) <- paste0("Gene-", 1:nrow(count_matrix))

# Generate example colInfo
col_info <- data.frame(cell_barcode = colnames(count_matrix))
col_info$batch <- sample(1:4, nrow(col_info), replace = TRUE)

# Create test EMSet
em_set <- newEMSet(list(counts = count_matrix), colInfo = col_info)

# Normalise by RLE
norm_set <- normaliseBatches(em_set)

}
