% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{normaliseBatches}
\alias{normaliseBatches}
\title{normaliseBatches}
\usage{
normaliseBatches(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} with cells from more than one batch.}
}
\value{
An \code{\linkS4class{EMSet}} with batch-normalised expression values.
}
\description{
Normalise counts to remove batch effects. This normalisation method is for
experiments where data from batches of different samples are combined without
undergoing library equalisation.
}
\details{
This step should be done prior to any filtering and normalisation between cells.
}
\examples{
# Load example EMSet
em_set <- ascend::raw_set

# Normalise batches
norm_set <- normaliseBatches(em_set)

}
