% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{NormaliseBatches}
\alias{NormaliseBatches}
\title{NormaliseBatches}
\usage{
NormaliseBatches(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet} with cells from more than one batch}
}
\value{
An \linkS4class{EMSet} with batch-normalised expression values
}
\description{
Normalise counts to remove batch effects. This normalisation method is for
experiments where data from batches of different samples are combined without
undergoing library equalisation.
}
\details{
This step should be done prior to any filtering and normalisation between cells.
}
\examples{
\dontrun{
batch_normalised_object <- NormaliseBatches(em.set)
}
}
