% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{normaliseByRLE}
\alias{normaliseByRLE}
\title{normaliseByRLE}
\usage{
normaliseByRLE(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} set that has undergone filtering. 
Please ensure spike-ins have been removed before using this function.}
}
\value{
An \code{\linkS4class{EMSet}} with normalised expression values.
}
\description{
Normalisation of expression between cells, by scaling to relative log 
expression (RLE). This method assumes all genes express a pseudo value higher 
than 0, and also assumes most genes are not differentially expressed. Only 
counts that are greater than zero are considered in this normalisation method.
}
\examples{
# Generate example matrix
count_matrix <- matrix(sample(0:1, 900, replace=TRUE),30,30)
colnames(count_matrix) <- paste0("Cell-", 1:ncol(count_matrix))
rownames(count_matrix) <- paste0("Gene-", 1:nrow(count_matrix))

# Generate example colInfo
col_info <- data.frame(cell_barcode = colnames(count_matrix))
col_info$batch <- sample(1:4, nrow(col_info), replace = TRUE)

# Create test EMSet
em_set <- newEMSet(assays = list(counts = count_matrix), colInfo = col_info)

# Normalise by RLE
norm_set <- normaliseByRLE(em_set)

}
