% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{NormaliseByRLE}
\alias{NormaliseByRLE}
\title{NormaliseByRLE}
\usage{
NormaliseByRLE(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} set that has undergone filtering. 
Please ensure spike-ins have been removed before using this function.}
}
\value{
An \code{\linkS4class{EMSet}} with normalised expression values.
}
\description{
Normalisation of expression between cells, by scaling to relative log 
expression (RLE). This method assumes all genes express a pseudo value higher 
than 0, and also assumes most genes are not differentially expressed. Only 
counts that are greater than zero are considered in this normalisation method.
}
\examples{
EMSet <- readRDS(system.file(package = "ascend", "extdata", "ExampleEMSet.rds"))
NormalisedEMSet <- NormaliseByRLE(EMSet)
}
