% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_normalisation.R
\name{NormaliseByRLE}
\alias{NormaliseByRLE}
\title{NormaliseByRLE}
\usage{
NormaliseByRLE(object)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object that has undergone filtering. Please ensure spike-ins have been removed before using this function.}
}
\description{
Normalisation of expression between cells, by scaling to relative log expression.
This method assumes all genes express a pseudo value higher than 0, and also assumes most genes
are not differentially expressed. Only counts that are greater than zero are considered in this normalisation method.
}
