% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{normWithinBatch}
\alias{normWithinBatch}
\title{normWithinBatch}
\usage{
normWithinBatch(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet}.}
}
\value{
An EMSet with values normalised within a single batch.
}
\description{
Normalise cells within each batch.
}
\examples{
# Generate example matrix
count_matrix <- matrix(sample(0:1, 900, replace=TRUE),30,30)
colnames(count_matrix) <- paste0("Cell-", 1:ncol(count_matrix))
rownames(count_matrix) <- paste0("Gene-", 1:nrow(count_matrix))

# Generate example colInfo
col_info <- data.frame(cell_barcode = colnames(count_matrix))
col_info$batch <- sample(1:4, nrow(col_info), replace = TRUE)

# Create test EMSet
em_set <- newEMSet(assays = list(counts = count_matrix), colInfo = col_info)

# Normalise within batches
norm_set <- normWithinBatch(em_set)

}
