% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{normWithinBatch}
\alias{normWithinBatch}
\title{normWithinBatch}
\usage{
normWithinBatch(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet}.}
}
\value{
An EMSet with values normalised within a single batch.
}
\description{
Normalise cells within each batch.
}
