% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_normalisation.R
\name{NormWithinBatch}
\alias{NormWithinBatch}
\title{NormWithinBatch}
\usage{
NormWithinBatch(batch.id, expression.matrix = NULL, cell.info = NULL)
}
\arguments{
\item{batch.id}{Batch identifier as stored in CellInformation.}

\item{expression.matrix}{Expression matrix that has not been batch-normalised.}

\item{cell.info}{Cell Information data frame.}
}
\value{
A list containing the normalisation factor and a batch-normalised 
expression matrix in sparse format.
}
\description{
Called by NormaliseBatches. Performs normalisation within a batch.
}
