% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotConsensusBars}
\alias{PlotConsensusBars}
\title{PlotConsensusBars}
\usage{
PlotConsensusBars(dendro, colors, rowLabels = NULL, rowWidths = NULL,
  rowText = NULL, rowTextAlignment = c("left", "center", "right"),
  rowTextIgnore = NULL, textPositions = NULL, addTextGuide = TRUE,
  cex.rowLabels = 1, cex.rowText = 0.8, ...)
}
\arguments{
\item{dendro}{hclust object generated by \code{\link{runCORE}}.}

\item{colors}{List of colours to use for the coloured bars.}

\item{rowLabels}{List of labels for colour bars.}

\item{rowWidths}{Width of the bars.}

\item{rowText}{Text to annotate bars.}

\item{rowTextAlignment}{Positioning of text - left, center or right.}

\item{rowTextIgnore}{Ignore row text.}

\item{textPositions}{Positoning of text.}

\item{addTextGuide}{TRUE or FALSE (Default).}

\item{cex.rowLabels}{1 (Default).}

\item{cex.rowText}{0.8 (Default).}

\item{...}{Additional arguments to pass onto function.}
}
\value{
Draws bars underneath dendrogram
}
\description{
Adapted from plotColorUnderTree found in the \pkg{WGCNA} package. This is
called by \code{\link{plotStabilityDendro}} to generate the consensus bars
beneath the dendrogram.
}
