% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotControlHist}
\alias{plotControlHist}
\title{plotControlHist}
\usage{
plotControlHist(object, control = NULL)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object.}

\item{control}{Name of control group to plot.}
}
\value{
A ggplot2 histogram
}
\description{
Generates a histogram of percentage of control per samples.
}
\examples{
# Load unprocessed EMSet
em_set <- ascend::raw_set

# Plot controls
control_histogram <- plotControlHist(em_set, control = "Mt")

}
