% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotControlPercentagesPerSample}
\alias{PlotControlPercentagesPerSample}
\title{PlotControlPercentagesPerSample}
\usage{
PlotControlPercentagesPerSample(object, control.name)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object}

\item{control.name}{Name of the control you would like to plot}
}
\description{
PlotControlPercentagesPerSample
}
