% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotDendrogram}
\alias{PlotDendrogram}
\title{PlotDendrogram}
\usage{
PlotDendrogram(object)
}
\arguments{
\item{object}{An \linkS4class{AEMSet} that has undergone clustering}
}
\description{
PlotDendrogram
}
