% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotDendrogram}
\alias{PlotDendrogram}
\title{PlotDendrogram}
\usage{
PlotDendrogram(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has undergone clustering.}
}
\value{
A dendrogram plotted by base R graphics
}
\description{
Generates a colour-labelled dendrogram to the device, using a clustered
\code{\linkS4class{EMSet}}.
}
\examples{
\dontrun{
# Open PDF file to output dendrogram to
pdf("Dendrogram.pdf", width= 4, height = 4)

# Plot dendrogram
PlotDendrogram(EMSet)

# Close PDF file
dev.off()
}

}
