% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotDendrogram}
\alias{PlotDendrogram}
\title{PlotDendrogram}
\usage{
PlotDendrogram(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has undergone clustering.}
}
\value{
A dendrogram plotted by base R graphics
}
\description{
Generates a colour-labelled dendrogram to the device, using a clustered
\code{\linkS4class{EMSet}}.
}
\examples{
# Load clustered EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", 
"ExampleClusteredEMSet.rds"))

# Plot dendrogram
PlotDendrogram(EMSet)

}
