% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotDendrogram}
\alias{plotDendrogram}
\title{plotDendrogram}
\usage{
plotDendrogram(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has undergone clustering.}
}
\value{
A dendrogram plotted by base R graphics
}
\description{
Generates a colour-labelled dendrogram to the device, using a clustered
\code{\linkS4class{EMSet}}.
}
\examples{
\dontrun{
plotDendrogram(em_set)
}
}
