% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotDEVolcano}
\alias{PlotDEVolcano}
\title{PlotDEVolcano}
\usage{
PlotDEVolcano(de.results, threshold = 0.005, l2fc = 2, labels = FALSE,
  label.size = 5)
}
\arguments{
\item{de.results}{Differential expression results generated by \code{\link{RunDifferentialExpression}} or \link{\link{RunClusterDiffExpression}}.}

\item{threshold}{Threshold to determine significant results by.}

\item{l2fc}{Threshold to determine significant Log2FoldChange by.}

\item{labels}{Display names of significant results on plot.}

\item{label.size}{Size of label text}
}
\description{
Produces a volcano plot featuring differential expression results.
}
