% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotGeneNumber}
\alias{plotGeneNumber}
\title{plotGeneNumber}
\usage{
plotGeneNumber(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet}}
}
\description{
Generates a histogram of the number of genes per cell.
}
\examples{
\dontrun{
gene_number_plot <- plotGeneNumber(EMSet)
}
}
