% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotGeneralQC}
\alias{PlotGeneralQC}
\title{PlotGeneralQC}
\usage{
PlotGeneralQC(object)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object}
}
\description{
This function generates the following plots:
}
