% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotGeneralQC}
\alias{PlotGeneralQC}
\title{PlotGeneralQC}
\usage{
PlotGeneralQC(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object}
}
\value{
A list of ggplot globs
}
\description{
This function generates the following plots:
\itemize{
\item{\strong{Library Size}: A histogram depicting library sizes across the dataset.}
}
}
