% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotLibrarySizesPerSample}
\alias{PlotLibrarySizesPerSample}
\title{PlotLibrarySizesPerSample}
\usage{
PlotLibrarySizesPerSample(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object}
}
\description{
Generates a violin-beeswarm plot of library sizes per sample. This is called 
by the \code{\link{PlotGeneralQC}} function.
}
