% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotMDS}
\alias{PlotMDS}
\title{PlotMDS}
\usage{
PlotMDS(object, PCA = FALSE, dim1 = 1, dim2 = 2, condition = NULL)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has undergone clustering with 
\code{\link{RunCORE}}.}

\item{PCA}{If true, use PCA-reduced matrix to generate MDS plot.}

\item{dim1}{Which dimension to plot on the x-axis.}

\item{dim2}{Which dimension to plot on the y-axis.}

\item{condition}{(Optional) Name of the condition in Cell Identifiers that 
describe a set of conditions you would like to colour cells by.}
}
\value{
A ggplot glob that contains a scatter plot.
}
\description{
Generates a Multi-Dimensional Scaling (MDS) plot.
}
\examples{
# Load clustered EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", 
"ExampleClusteredEMSet.rds"))

# Load ggplot to add title to MDS plot
library(ggplot2)

# Plot MDS
my_mds_plot <- PlotMDS(EMSet, PCA = TRUE, dim1 = 3, dim2 = 4, 
condition = "cluster") + ggtitle("My MDS Plot")

}
