% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotMDS}
\alias{PlotMDS}
\title{PlotMDS}
\usage{
PlotMDS(object, PCA = FALSE, dim1 = 1, dim2 = 2, condition = NULL)
}
\description{
Generates a Multi-Dimensional Scaling (MDS) plot.
}
\details{
@param object An \linkS4class{AEMSet} that has undergone clustering with \code{FindOptimalClusters}.
 @param PCA If true, use PCA-reduced matrix to generate MDS plot
 @param dim1 Which dimension to plot on the x-axis
 @param dim2 Which dimension to plot on the y-axis
 @param condition (Optional) Name of the condition in Cell Identifiers that describe a set of conditions you would like to colour cells by
 @export
}
