% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotMDS}
\alias{plotMDS}
\title{plotMDS}
\usage{
plotMDS(object, Dim1 = 1, Dim2 = 2, group = NULL, density = FALSE)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}}.}

\item{Dim1}{Dimension to plot on the x-axis.}

\item{Dim2}{Dimension to plot on the y-axis.}

\item{group}{(Optional) Name of the column in colInfo that 
describe a set of conditions you would like to colour cells by.}

\item{density}{Vary alpha by density (Default: FALSE)}
}
\value{
A ggplot glob that contains a scatter plot.
}
\description{
Generates a 2D MDS plot that can use the following inputs stored in an
EMSet:
}
\details{
1. Distance matrix generated by \link{runCORE}.
2. PCA matrix generated by \code{\link{runPCA}}.
3. Normalised, log-transformed count matrix stored in logcounts.
}
\examples{
\dontrun{
# Normalise counts
EMSet <- normaliseByRLE(EMSet)

# Remove controls
EMSet <- excludeControl(EMSet, control = c("Mt", "Rb"))
# Dimensionality reduction
EMSet <- runPCA(EMSet)

# Clustering
EMSet <- runCORE(EMSet)

# Generate MDS
mds_plot <- plotMDS(EMSet, Dim1 = 1, Dim2 = 2, group = "cluster")

}
}
