% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotNormalisationQC}
\alias{PlotNormalisationQC}
\title{PlotNormalisationQC}
\usage{
PlotNormalisationQC(original = NULL, normalised = NULL, gene = NULL)
}
\arguments{
\item{original}{An un-normalised \linkS4class{AEMSet}}

\item{normalised}{A normalised \linkS4class{AEMSet}}

\item{gene}{Gene to plot expression levels for (optional)}
}
\description{
Generates a series of plots comparing an un-normalised and a normalised \linkS4class{AEMSet}.
}
