% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotNormalisationQC}
\alias{PlotNormalisationQC}
\title{PlotNormalisationQC}
\usage{
PlotNormalisationQC(original = NULL, normalised = NULL,
  gene.list = list())
}
\arguments{
\item{original}{An un-normalised \linkS4class{EMSet}}

\item{normalised}{A normalised \linkS4class{EMSet}}

\item{gene.list}{OPTIONAL: A list of genes to plot expression levels for. If not defined, ASCEND will choose a gene at random}
}
\description{
Generates a series of plots comparing an un-normalised and a normalised \linkS4class{EMSet}.
}
