% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotPCA}
\alias{PlotPCA}
\title{PlotPCA}
\usage{
PlotPCA(object, dim1 = 1, dim2 = 2, condition.list = list())
}
\arguments{
\item{object}{An \linkS4class{AEMSet} object that has undergone PCA}

\item{dim1}{Principal component to plot on the x-axis}

\item{dim2}{Principal component to plot on the y-axis}

\item{condition.list}{(Optional) A list of barcodes and associated conditions to colour points by}
}
\description{
Plot two principal components (PCs) on a scatter plot.
}
