% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA}
\usage{
plotPCA(object, PCX = 1, PCY = 2, group = NULL)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object that has undergone PCA.}

\item{PCX}{Principal component to plot on the x-axis.}

\item{PCY}{Principal component to plot on the y-axis.}

\item{group}{(Optional) Name of the column in colInfo that 
describe a set of conditions you would like to colour cells by.}
}
\value{
A ggplot glob that contains a scatter plot.
}
\description{
Plot two principal components (PCs) on a scatter plot. This plot corresponds
more closely to the distance between points and therefore is good to use
to review the effectiveness of clustering by the CORE algorithm.
}
