% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotPCA}
\alias{PlotPCA}
\title{PlotPCA}
\usage{
PlotPCA(object, dim1 = 1, dim2 = 2, condition = NULL)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object that has undergone PCA}

\item{dim1}{Principal component to plot on the x-axis}

\item{dim2}{Principal component to plot on the y-axis}

\item{condition}{(Optional) Name of the column in CellIdentifiers that describe a set of conditions you would like to colour cells by}
}
\description{
Plot two principal components (PCs) on a scatter plot.
}
