% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotPCAVariance}
\alias{PlotPCAVariance}
\title{PlotPCAVariance}
\usage{
PlotPCAVariance(object, n = 100)
}
\arguments{
\item{object}{An \linkS4class{EMSet} that has undergone PCA.}

\item{n}{Number of PCA values on the plot}
}
\description{
Plots PCA Percent Variance.
}
