% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotPCAVariance}
\alias{PlotPCAVariance}
\title{PlotPCAVariance}
\usage{
PlotPCAVariance(object, n = 100)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has undergone PCA.}

\item{n}{Number of PCs to view on the plot.}
}
\value{
A ggplot2 glob that contains a scatter qplot.
}
\description{
Generates a scree plot. The principal components (PCs) are sorted from 
largest percent variance to the smallest percent variance. This plot is
for determining the optimal number of PCs to retain for further analysis.
}
\examples{
# Load clustered EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", 
"ExampleClusteredEMSet.rds"))

# Plot PCA Variance
pca_variance <- PlotPCAVariance(EMSet, n = 20)

}
