% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_utilities.R
\name{PlotPDF}
\alias{PlotPDF}
\title{PlotPDF}
\usage{
PlotPDF(plot.list, filename = NULL, width = 11, height = 8.5)
}
\arguments{
\item{plot.list}{A list of plots generated by \pkg{ASCEND}}

\item{filename}{Name of the file you would like to output to}

\item{width}{Width of the PDF}

\item{height}{Height of the PDF}
}
\description{
Takes a list of plots generated by \pkg{ASCEND} and prints them to PDF.
}
