% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotSmoothScatter}
\alias{plotSmoothScatter}
\title{plotSmoothScatter}
\usage{
plotSmoothScatter(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet}.}
}
\value{
A ggplot2 glob containing the histogram.
}
\description{
Generates a smooth scatter of of average counts for each gene per cell.
}
\examples{
# Load EMSet
EMSet <- ascend::raw_set

smooth_scatter <- plotSmoothScatter(EMSet)

}
