% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotStability}
\alias{plotStability}
\title{plotStability}
\usage{
plotStability(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object that has undergone clustering.}
}
\value{
A line graph generated by ggplot2's geom_line function
}
\description{
Plots Stability, Consecutive RI and Rand Index. This can be used to determine
the optimal resolution of the clustering results.
}
\examples{
\dontrun{
stability_plot <- plotStability(em_set)
}
}
