% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotStability}
\alias{PlotStability}
\title{PlotStability}
\usage{
PlotStability(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object that has undergone clustering}
}
\description{
Plots Stability, Consecutive RI and Rand Index. This can be used to determine
the optimal resolution of the clustering results.
}
