% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotStabilityDendro}
\alias{plotStabilityDendro}
\title{plotStabilityDendro}
\usage{
plotStabilityDendro(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has undergone clustering.}
}
\value{
A dendrogram with coloured bars generated by R base graphics
}
\description{
Generate a dendrogram with coloured bars representing each clustering result
below. This function is derived from the plotDendroAndColors function found
in the \pkg{WGCNA} package.
}
\examples{
\dontrun{
pdf("MyStabilityDendrogram.pdf")
#' # Load clustered EMSet
# Plot stability dendrogram
PlotStabilityDendro(EMSet)
dev.off()
}
}
