% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{PlotStabilityDendro}
\alias{PlotStabilityDendro}
\title{PlotStabilityDendro}
\usage{
PlotStabilityDendro(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has undergone clustering.}
}
\value{
A dendrogram with coloured bars generated by R base graphics
}
\description{
Generate a dendrogram with coloured bars representing each clustering result
below. This function is derived from the plotDendroAndColors function found
in the \pkg{WGCNA} package.
}
\examples{
\dontrun{
pdf("MyStabilityDendrogram.pdf")
}
#' # Load clustered EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", 
"ExampleClusteredEMSet.rds"))

# Plot stability dendrogram
PlotStabilityDendro(EMSet)
\dontrun{
dev.off()
}
}
