% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotTopGenesPerSample}
\alias{plotTopGenesPerSample}
\title{plotTopGenesPerSample}
\usage{
plotTopGenesPerSample(object)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object.}
}
\value{
A ggplot2 object containing a violin-beeswarm plot
}
\description{
Generates a violin-beeswarm plot of top gene counts and contribution to to 
a cell's total expression.
}
\examples{
\dontrun{
top_genes_violin <- plotTopGenesPerSample(EMSet)
}
}
