% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotTSNE}
\alias{plotTSNE}
\title{plotTSNE}
\usage{
plotTSNE(object, Dim1 = 1, Dim2 = 2, group = NULL, density = FALSE)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object that has undergone PCA.}

\item{Dim1}{Dimension to plot on the x-axis.}

\item{Dim2}{Dimension to plot on the y-axis.}

\item{group}{(Optional) Name of the column in colInfo that 
describe a set of conditions you would like to colour cells by.}

\item{density}{Vary alpha by density (Default: FALSE)}
}
\value{
A ggplot glob that contains a scatter plot.
}
\description{
Generates a 2D TSNE plot using a TSNE matrix generated by the 
\code{\link[ascend:runTSNE-EMSet-method]{runTSNE}} function.
}
\examples{
\dontrun{
# Normalise counts
EMSet <- normaliseByRLE(EMSet)

# Remove controls
EMSet <- excludeControl(EMSet, control = c("Mt", "Rb"))

# Dimensionality reduction with PCA
EMSet <- runPCA(EMSet)

# Clustering
EMSet <- runCORE(EMSet)

# Dimensionality reduction with t-SNE
EMSet <- runTSNE(EMSet, seed = 1)

# Plot with t-SNE 
tsne_plot <- plotTSNE(EMSet, Dim1 = 1, Dim2 = 2, group = "cluster")

}
}
