% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotTSNE}
\alias{plotTSNE}
\title{plotTSNE}
\usage{
plotTSNE(object, Dim1 = 1, Dim2 = 2, group = NULL)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} object that has undergone PCA.}

\item{Dim1}{Dimension to plot on the x-axis.}

\item{Dim2}{Dimension to plot on the y-axis.}

\item{group}{(Optional) Name of the column in colInfo that 
describe a set of conditions you would like to colour cells by.}
}
\value{
A ggplot glob that contains a scatter plot.
}
\description{
Generates a 2D TSNE plot using a TSNE matrix generated by the 
\code{\link{runTSNE}} function.
}
