% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_plots.R
\name{PlotTSNE}
\alias{PlotTSNE}
\title{PlotTSNE}
\usage{
PlotTSNE(object, PCA = TRUE, column = NULL, colour = NULL, seed = 0,
  perplexity = 30, theta = 0.5)
}
\arguments{
\item{object}{An \linkS4class{AEMSet}.}

\item{PCA}{Set to FALSE to not use PCA-reduced values}

\item{column}{(Optional) Name of the column in CellIdentifiers that describe a set of conditions you would like to colour cells by}

\item{colour}{(Optional) Name of the column in CellIdentifiers that contains colours you wish each cell and condition to be associated with.}

\item{seed}{(Optional) Set to a specific value for reproducible TSNE plots}

\item{perplexity}{(Optional) Numeric; perplexity parameter}

\item{theta}{(Optional) Numeric; Speed/accuracy trade-off (increase for less accuracy)}
}
\description{
Generates a 2D TSNE plot.
}
