% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_plots.R
\name{plotVolcano}
\alias{plotVolcano}
\title{plotVolcano}
\usage{
plotVolcano(x, threshold = 0.005, l2fc = 2, labels = FALSE,
  label.size = 5, check.overlap = TRUE)
}
\arguments{
\item{x}{Differential expression results generated by 
\code{\link{runDESeq}}.}

\item{threshold}{Threshold to determine significant results by (Default: 5e-3).}

\item{l2fc}{Threshold to determine significant Log2FoldChange by (Default: 2).}

\item{labels}{Display names of significant results on plot (Default: FALSE).}

\item{label.size}{Size of label text (Default: 5).}

\item{check.overlap}{If labels = TRUE, remove overlapping labels}
}
\value{
A scatter plot rendered by ggplot2's geom_point.
}
\description{
Produces a volcano plot featuring differential expression results.
}
\examples{
\dontrun{
library(ggplot2)
my_volcano_plot <- PlotDEVolcano(de_result_df, threshold = 5e-3, l2fc = 2) +
ggtitle("My Volcano Plot")
}

}
