% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_diffexpression.R
\name{PrepareCountData}
\alias{PrepareCountData}
\title{PrepareCountData}
\usage{
PrepareCountData(x)
}
\description{
Called by RunDiffExpression. This chunks up the expression matrix to feed into DESeq.
}
