% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_diffexpression.R
\name{ProcessDEResults}
\alias{ProcessDEResults}
\title{ProcessDEREsults}
\usage{
ProcessDEResults(output.list)
}
\description{
Called by RunDiffExpression. Compiles the resultant data into data frames and converts
the results to absolute fold change.
}
