% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_cellranger.R
\name{ProcessGeneReference}
\alias{ProcessGeneReference}
\title{ProcessGeneReference}
\usage{
ProcessGeneReference(gene.obj)
}
\arguments{
\item{gene.obj}{Gene information stored in a data frame imported from Cell
Ranger output.}
}
\value{
A dataframe containing gene_symbols in the first column.
}
\description{
Called by \code{\link{BuildCellRangerDataset}} to parse genes.tsv into a data
frame.
}
