% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_getters.R, R/ascend_setters.R
\docType{methods}
\name{regcounts}
\alias{regcounts}
\alias{regcounts,EMSet-method}
\alias{regcounts<-}
\alias{regcounts<-,EMSet-method}
\title{regcounts}
\usage{
regcounts(x, ...)

\S4method{regcounts}{EMSet}(x)

regcounts(x, ...) <- value

\S4method{regcounts}{EMSet}(x) <- value
}
\arguments{
\item{x}{\linkS4class{EMSet}}

\item{...}{...}

\item{value}{A matrix with confounding factors regressed out}
}
\value{
Regressed count matrix
}
\description{
Convenience method to access regressed counts from an \linkS4class{EMSet}.
}
\examples{
# Load em_set
em_set <- ascend::analyzed_set

# Regress out CDK4
reg_set <- regressConfoundingFactors(em_set, candidate.genes = "CDK4")

# Access regressed matrix
reg_counts <- regcounts(reg_set)

}
