% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_confoundingfactors.R
\name{regressConfoundingFactors}
\alias{regressConfoundingFactors}
\title{regressConfoundingFactors}
\usage{
regressConfoundingFactors(object, candidate.genes = c())
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has been normalised.}

\item{candidate.genes}{A list of genes you wish to regress from the dataset.
Refer to the vignette on how to choose genes for regression.}
}
\value{
An \code{\linkS4class{EMSet}} with confounding factors regressed 
from the expression values.
}
\description{
This function generates a scaled regression matrix based on candidate genes
supplied by the user. This function should be used after normalisation.
}
\examples{
# Load example EMSet
em_set <- ascend::analyzed_set

# Define genes to regress
genes <- c("CDK4","CCND1")

regressedSet <- regressConfoundingFactors(em_set, candidate.genes = genes)

}
