% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_confoundingfactors.R
\name{RegressConfoundingFactors}
\alias{RegressConfoundingFactors}
\title{RegressConfoundingFactors}
\usage{
RegressConfoundingFactors(object, candidate.genes = list())
}
\arguments{
\item{candidate.genes}{A list of genes you wish to regress from the dataset.
Refer to the vignette on how to choose genes for regression.}

\item{expression.matrix}{Expression matrix in data frame format. You may load
it from a EMSet object with the GetExpressionMatrix function.}
}
\description{
This function generates a scaled regression matrix based on candidate genes
supplied by the user. This function is best used after normalisation.
}
