% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_confoundingfactors.R
\name{regressGene}
\alias{regressGene}
\title{regressGene}
\usage{
regressGene(x, covariate.matrix = NULL, expression.matrix = NULL)
}
\arguments{
\item{x}{Gene to regress residuals from.}

\item{covariate.matrix}{Covariate matrix generated by RemoveConfoundingFactors.}

\item{expression.matrix}{A transposed expression matrix generated by 
RemoveConfoundingFactors.}
}
\value{
List of values generated from regression
}
\description{
Called by RemoveConfoundingFactors. This step extracts residuals from the 
supplied gene.
}
