% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\name{ReplaceCellInfo}
\alias{ReplaceCellInfo}
\title{ReplaceCellInfo}
\usage{
ReplaceCellInfo(object, cell.info)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object.}

\item{cell.info}{A data frame containing cell information. The first column must comprise of cell identifiers that match the cell identifiers used in the expression matrix. The second column, while optional - should contain batch information.}
}
\description{
Can be called by the user or by a filtering function. Updates Cell Information in a \linkS4class{EMSet} object, replacing the old data frame with a new one.
}
