% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\docType{methods}
\name{ReplaceExpressionMatrix,EMSet-method}
\alias{ReplaceExpressionMatrix,EMSet-method}
\title{ReplaceExpressionMatrix}
\usage{
\S4method{ReplaceExpressionMatrix}{EMSet}(object, expression.matrix)
}
\arguments{
\item{object}{The \linkS4class{EMSet} you would like to update.}

\item{expression.matrix}{Expression matrix in matrix form}
}
\value{
An \linkS4class{EMSet} with an updated expression matrix
}
\description{
Replace the expression matrix in a \linkS4class{EMSet} with a new expression 
matrix and re-calculate its metrics.
}
\examples{
# Load an example EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", "ExampleEMSet.rds"))
    
# Load an example matrix
new_matrix <- readRDS(system.file(package = "ascend", "extdata", "ExampleMatrix.rds"))

# Subset cells from it
subset_matrix <- new_matrix[ , 1:100]
subset_matrix <- as.matrix(subset_matrix)

# Replace it with another expression matrix
UpdatedEMSet <- ReplaceExpressionMatrix(EMSet, subset_matrix)

}
