% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{ReplaceExpressionMatrix}
\alias{ReplaceExpressionMatrix}
\title{ReplaceExpressionMatrix}
\usage{
ReplaceExpressionMatrix(object, expression.matrix)
}
\arguments{
\item{object}{The \linkS4class{EMSet} you would like to update.}

\item{expression.matrix}{Expression matrix in matrix form}
}
\description{
Definition for \code{\link{ReplaceExpressionMatrix}} function.
}
