% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\name{ReplaceExpressionMatrix}
\alias{ReplaceExpressionMatrix}
\title{ReplaceExpressionMatrix}
\usage{
ReplaceExpressionMatrix(x, object)
}
\arguments{
\item{x}{Expression matrix in matrix or data.frame}

\item{object}{The \linkS4class{AEMSet} you would like to update}
}
\description{
Replace the expression matrix in a \linkS4class{AEMSet} with a new expression matrix and re-calculate its metrics.
}
