% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{ReplaceGeneInfo}
\alias{ReplaceGeneInfo}
\title{ReplaceGeneInfo}
\usage{
ReplaceGeneInfo(object, gene.info)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object.}

\item{gene.info}{A data frame containing cell information. The first column 
must comprise of gene identifiers that match the gene identifiers used in the 
expression matrix. The remaining columns can contain anything.}
}
\value{
An \linkS4class{EMSet} with updated gene information.
}
\description{
Can be called by the user or by a filtering function. Updates Gene 
Information in a \linkS4class{EMSet} object, replacing the old data frame 
with a new one.
}
\examples{
# Load an EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", "ExampleEMSet.rds"))
    
# Retrieve "Gene Information" dataframe from EMSet
old_gene_info <- GetGeneInfo(EMSet)

# Update gene information
new_gene_info <- old_gene_info
new_gene_info$updated <- TRUE

# Replace Cell Information
updatedEMSet <- ReplaceGeneInfo(EMSet, new_gene_info)

}
