% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_clustering.R
\name{RetrieveCluster}
\alias{RetrieveCluster}
\title{RetrieveCluster}
\usage{
RetrieveCluster(height.list, hclust.obj = NULL, distance.matrix = NULL)
}
\arguments{
\item{height.list}{List of resolutions to cut at}

\item{hclust.obj}{Hclust object generated from the distance matrix}

\item{distance.matrix}{Distance matrix generated by RunCORE from PCA matrix}
}
\value{
List of clusters
}
\description{
Generates clusters using dynamicTreeCut's unsupervised cut setting.
}
