% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_setters.R, R/ascend_getters.R
\docType{methods}
\name{rowInfo<-}
\alias{rowInfo<-}
\alias{rowInfo<-,EMSet-method}
\alias{rowInfo}
\alias{rowInfo,EMSet-method}
\title{rowInfo}
\usage{
rowInfo(x, ...) <- value

\S4method{rowInfo}{EMSet}(x) <- value

rowInfo(x, ...)

\S4method{rowInfo}{EMSet}(x, withDimnames = TRUE)
}
\arguments{
\item{x}{\linkS4class{EMSet}}

\item{...}{...}

\item{value}{DataFrame to store in rowInfo slot.}

\item{withDimnames}{TRUE or FALSE}
}
\value{
rowInfo DataFrame
}
\description{
Access gene information stored in rowInfo.
}
\examples{

# Load EMSet
em_set <- ascend::raw_set

# Retrieve colInfo
row_info <- rowInfo(em_set)

# Set colInfo
rowInfo(em_set) <- row_info

}
