% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_diffexpression.R
\name{RunDESeq}
\alias{RunDESeq}
\title{RunDESeq}
\usage{
RunDESeq(data, condition.list = list(), condition.a = NULL,
  condition.b = NULL, fitType = NULL, method = NULL)
}
\arguments{
\item{data}{Chunk of count matrix}

\item{condition.list}{List of conditions to test}

\item{condition.a}{Condition A}

\item{condition.b}{Condition B}

\item{fitType}{Type of fit to use with \pkg{DESeq}}

\item{method}{Method to use with \pkg{DESeq}}
}
\value{
A dataframe containing DESeq results
}
\description{
Called by \code{\link{RunDiffExpression}} to run in parallel. This performs 
the differential expression part.
}
