% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_DESeq.R
\name{runDESeq}
\alias{runDESeq}
\title{runDESeq}
\usage{
runDESeq(object, group = NULL, condition.a = NULL,
  condition.b = NULL, ngenes = NULL, fitType = c("parametric",
  "local"), method = c("pooled", "pooled-CR", "per-condition", "blind"),
  parallel = TRUE)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object that has undergone 
filtering and normalisation.}

\item{group}{Name of the column in the colInfo dataframe where you have
defined the conditions you would like to test. eg cluster to compare clusters
identified by \code{\link[ascend:runCORE]{runCORE}}.}

\item{condition.a}{Condition of the group you want to use as the baseline.}

\item{condition.b}{Conditions of the group you want to compare to the baseline.}

\item{ngenes}{Perform differential expression analysis using top number of genes.
If omitted, this function will run analysis on ALL genes.}

\item{fitType}{Method used to fit a dispersion-mean relation by \pkg{DESeq}. 
Options: parametric, local (Default).}

\item{method}{Method used by \pkg{DESeq} to compute emperical dispersion.
Options: pooled, pooled-CR, per-condition (Default), blind.}

\item{parallel}{Run DESeq through parallelised wrapper (Default: TRUE)}
}
\value{
A data frame containing \pkg{DESeq} results
}
\description{
This wrapper runs differential expression analysis via the \pkg{DESeq} 
package.
}
\examples{
\dontrun{
library(DESeq)
cluster1_vs_others <- runDESeq(EMSet, group = "cluster", condition.a = "1",
condition.b = c("2", "3"), ngenes = 1500, fitType = "local", method = "per-condition")
}

}
