% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_diffexpression.R
\name{RunDiffExpression}
\alias{RunDiffExpression}
\title{RunDiffExpression}
\usage{
RunDiffExpression(object, conditions = NULL, condition.a = NULL,
  condition.b = NULL, fitType = c("parametric", "local"),
  method = c("pooled", "pooled-CR", "per-condition", "blind"))
}
\arguments{
\item{object}{A \linkS4class{EMSet} object that has undergone filtering and 
normalisation.}

\item{conditions}{Name of the column in the CellInformations lot where you have
defined the conditions you would like to test. eg cluster to compare clusters
identified by RunCORE.}

\item{condition.a}{Condition of the group you want to use as the baseline}

\item{condition.b}{Conditions of the group you want to compare to the baseline.}

\item{fitType}{Method used to fit a dispersion-mean relation by \pkg{DESeq}. 
Options: parametric, local (Default)}

\item{method}{Method used by \pkg{DESeq} to compute emperical dispersion.
Options: pooled, pooled-CR, per-condition (Default), blind}
}
\description{
Compare the differential expression of genes between cells of different conditions.
}
