% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_diffexpression.R
\name{RunDiffExpression}
\alias{RunDiffExpression}
\title{RunDiffExpression}
\usage{
RunDiffExpression(object, column = NULL, conditions = NULL)
}
\arguments{
\item{object}{A \linkS4class{EMSet} object that has undergone clustering
with the \code{\link{RunCORE}} function.}

\item{column}{Name of the column in the CellInformations lot where you have
defined the conditions you would like to test. eg cluster to compare clusters
identified by RunCORE.}

\item{conditions}{List of conditions you want to test, in the order you would
like to run them in. This list of terms should match those used in your selected column.}
}
\description{
Compare the differential expression of genes in each cluster versus other clusters.
}
