% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_diffexpression.R
\name{runDiffExpression}
\alias{runDiffExpression}
\title{runDiffExpression}
\usage{
runDiffExpression(object, group = NULL, condition.a = NULL,
  condition.b = NULL, subsampling = FALSE, ngenes = NULL)
}
\arguments{
\item{object}{An \linkS4class{EMSet}.}

\item{group}{A column in colInfo that contains a vector of conditions.}

\item{condition.a}{Condition(s) of first group of cells.}

\item{condition.b}{Condition(s) of second group of cells.}

\item{subsampling}{TRUE or FALSE (Default). Whether or not to subsample 
from larger group of cells if cell populations are uneven.}

\item{ngenes}{Test this number of the most variable genes in the dataset.}
}
\description{
This function uses a combined Likelihood-Ratio Test (LRT) for 
discrete/continuous models to examine differentially expressed genes, on a 
gene-gene level. This method was adapted from the method published by [McDavid
et al. 2013](https://doi.org/10.1093/bioinformatics/bts714).
}
\examples{
# Load example EMSet
em_set <- ascend::analyzed_set

# Compare cluster 1 vs cluster 2
cluster1_vs_cluster2 <- runDiffExpression(em_set, group = "cluster",
condition.a = 1, condition.b = 2, subsampling = FALSE, ngenes = 1500)

}
