% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_dimreduction.R
\docType{methods}
\name{runPCA}
\alias{runPCA}
\alias{runPCA,EMSet-method}
\title{runPCA}
\usage{
runPCA(object, ..., ngenes, scaling)

\S4method{runPCA}{EMSet}(object, ..., ngenes = 1500, scaling = TRUE)
}
\arguments{
\item{object}{An \linkS4class{EMSet} that has undergone filtering and 
normalisation.}

\item{...}{Additional arguments to pass to irlba}

\item{ngenes}{The top number of genes you would like to perform the reduction 
by. Default: 1500.}

\item{scaling}{Boolean - set to FALSE if you do not want to scale your values. 
Default: TRUE.}
}
\value{
An \linkS4class{EMSet} with a PCA-reduced matrix stored under PCA in
the reducedDims slot
}
\description{
Reduce the dimensions of an expression matrix stored in an 
\linkS4class{EMSet} based on the most variable genes. Datasets must
be reduced prior to clustering analysis as it is used to construct the 
distance matrix.
}
\examples{
# Load example EMSet
em_set <- ascend::analyzed_set

# Run PCA
em_set <- runPCA(em_set, scaling = TRUE, ngenes = 1500)


}
