% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_dimreduction.R
\name{RunPCA}
\alias{RunPCA}
\title{RunPCA}
\usage{
RunPCA(object, ngenes = 1500, scaling = TRUE)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has undergone filtering and 
normalisation.}

\item{ngenes}{The top number of genes you would like to perform the reduction by. 
Default: 1500}

\item{scaling}{Boolean - set to FALSE if you do not want to scale your values. 
Default: TRUE}
}
\value{
An \code{\linkS4class{EMSet}} with a PCA-reduced matrix stored in the PCA
slot.
}
\description{
Reduce the dimensions of an expression matrix stored in an 
\code{\linkS4class{EMSet}}. This should be used prior to clustering.
}
\examples{
\dontrun{
pca_set <- RunPCA(em.set)
}
}
