% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_dimreduction.R
\name{RunTSNE}
\alias{RunTSNE}
\title{RunTSNE}
\usage{
RunTSNE(object, PCA = FALSE, dimensions = 2, seed = 0, perplexity = 30,
  theta = 0.5)
}
\arguments{
\item{object}{An expression matrix or a PCA-reduced matrix.}

\item{PCA}{Set this PCA flag to true if the object is a PCA-reduced matrix.}

\item{dimensions}{Number of dimensions you would like to reduce to}

\item{seed}{(Optional) Set to a specific value for reproducible TSNE plots}

\item{perplexity}{(Optional) Numeric; perplexity parameter}

\item{theta}{(Optional) Numeric; Speed/accuracy trade-off (increase for less accuracy)}
}
\description{
RunTSNE
}
