% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_dimreduction.R
\docType{methods}
\name{runTSNE}
\alias{runTSNE}
\alias{runTSNE,EMSet-method}
\title{runTSNE}
\usage{
runTSNE(object, ..., PCA, dims, seed, perplexity, theta)

\S4method{runTSNE}{EMSet}(object, ..., PCA = FALSE, dims = 2,
  seed = 0, perplexity = 30, theta = 0.5)
}
\arguments{
\item{object}{An expression matrix or a PCA-reduced matrix.}

\item{...}{Additional arguments to pass on to \code{\link[Rtsne]{Rtsne}}
function.}

\item{PCA}{Set this PCA flag to TRUE if the object is a PCA-reduced matrix. 
Default: FALSE.}

\item{dims}{Number of dimensions you would like to reduce to. 
Default: 2.}

\item{seed}{(Optional) Set to a specific value for reproducible TSNE plots. 
Default: 0.}

\item{perplexity}{(Optional) Numeric; perplexity parameter. Default: 30.}

\item{theta}{(Optional) Numeric; Speed/accuracy trade-off. 
(increase for less accuracy). Default: 0.5.}
}
\value{
A dataframe containing expression data for each cell reduced to 
selected number of dimensions.
}
\description{
Wrapper for the \code{\link[Rtsne]{Rtsne}} function. Users may call this 
directly or call it through the \code{\link{plotTSNE}} function and also pass
additional arguments related to the \code{\link[Rtsne]{Rtsne}} function.
}
\examples{
# Load example EMSet
em_set <- ascend::analyzed_set

# Run tSNE
em_set <- runTSNE(em_set, PCA = TRUE, dims = 2, seed = 1)

}
