% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_scran.R
\name{SCESet2EMSet}
\alias{SCESet2EMSet}
\title{SCESet2EMset}
\usage{
SCESet2EMSet(SCESet, EMSet)
}
\arguments{
\item{SCESet}{A SCESet from \pkg{scater}.}

\item{EMSet}{An \code{\linkS4class{EMSet}} to load data from.}
}
\value{
An \code{\linkS4class{EMSet}} with data retrieved from a SCESet.
}
\description{
Loads data from a SCESet to a pre-existing 
\linkS4class{EMSet} object.
}
