% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_scran.R
\name{SCESetnormalise}
\alias{SCESetnormalise}
\title{SCESetnormalise}
\usage{
SCESetnormalise(sce.set, em.set, quickCluster = FALSE)
}
\arguments{
\item{sce.set}{A \pkg{SingleCellExperiment} object}

\item{em.set}{The \code{\linkS4class{EMSet}} to load the SCESet into}

\item{quickCluster}{Whether or not to use quickCluster (Default: FALSE)}
}
\value{
A normalised \code{\linkS4class{EMSet}} object
}
\description{
Called by \code{\link{scranNormalise}} - runs normalisation on a 
\pkg{SingleCellExperiment} object and converts it back to an 
\linkS4class{EMSet}.
}
