% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_scran.R
\name{scranNormalise}
\alias{scranNormalise}
\title{scranNormalise}
\usage{
scranNormalise(object, quickCluster = FALSE, min_mean = 1e-05)
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}} that has not undergone 
normalisation.}

\item{quickCluster}{Use scran's quickCluster method (TRUE) or  use randomly-
assigned groups (FALSE, Default).}

\item{min_mean}{Threshold for average counts. This argument is for the 
\code{\link[scran]{computeSumFactors}} function from \pkg{scran}. The value 
of 1 is recommended for read count data, while the default value of 1e-5 is 
best for UMI data. This argument is only used for newer versions of 
\pkg{scran}.}
}
\value{
An \code{\linkS4class{EMSet}} with an expression matrix with counts 
normalised by \code{\link[scater]{normalize}} function.
}
\description{
Normalise an \code{\linkS4class{EMSet}} with \pkg{scran}'s deconvolution 
method by Lun et al. 2016.
}
\details{
Users may choose to run computeSumFactors using either preset 
group sizes (40,60, 80 and 100) or quickCluster. For datasets with over
20,000 cells - it is recommended you use quickCluster.
}
