% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_normalisation.R
\name{scranNormalise}
\alias{scranNormalise}
\title{scranNormalise}
\usage{
scranNormalise(object)
}
\arguments{
\item{object}{An \linkS4class{AEMSet} that has not undergone normalisation.}
}
\description{
Normalise an \linkS4class{AEMSet} with \pkg{scran}'s deconvolution method by Lun et al. 2016.
}
