% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetmethods.R
\name{SubsetBatch}
\alias{SubsetBatch}
\title{SubsetBatch}
\usage{
SubsetBatch(object, batches)
}
\arguments{
\item{object}{An \linkS4class{EMSet}}

\item{batches}{Name or number of the batch(es) you would like to subset}
}
\value{
An \linkS4class{EMSet} containing only this batch
}
\description{
Subset a specific batch from an \linkS4class{EMSet}.
}
\examples{
\dontrun{
sample1_emset <- SubsetBatch(em.set, batches = c("1", "2", "3"))
}
}
