% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_subsetmethods.R
\name{SubsetBatch}
\alias{SubsetBatch}
\title{SubsetBatch}
\usage{
SubsetBatch(object, batches)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object}

\item{batches}{Name or number of the batch(es) you would like to subset.}
}
\value{
An \linkS4class{AEMSet} containing only this batch.
}
\description{
Subset a specific batch from a \linkS4class{AEMSet} object. This data is already normalised, but if you wish to recluster the data, you will need to use the RunPCA function again.
}
