% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_objects.R
\name{SubsetBatch}
\alias{SubsetBatch}
\title{SubsetBatch}
\usage{
SubsetBatch(object, x)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object}

\item{x}{Name or number of the batch you would like to subset.}
}
\description{
Subset a specific batch from a \linkS4class{AEMSet} object.
}
