% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetmethods.R
\docType{methods}
\name{SubsetCells,EMSet-method}
\alias{SubsetCells,EMSet-method}
\title{SubsetCells}
\usage{
\S4method{SubsetCells}{EMSet}(object, cell.barcodes = c())
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}}.}

\item{cell.barcodes}{A list of cell identifiers to subset from the 
\linkS4class{EMSet}.}
}
\value{
An \code{\linkS4class{EMSet}}.
}
\description{
Subset cells in the supplied list from an \code{\linkS4class{EMSet}}.
}
\examples{
# Load example EMSet
em.set <- readRDS(system.file(package = "ascend", "extdata", "ExampleEMSet.rds"))

# Randomly subsample cells from EMSet by choosing random barcodes from the EMSet
cell.info <- GetCellInfo(em.set)
barcode_list <- cell.info[, 1]
random_barcodes <- sample(barcode_list, 50, replace = FALSE)

# Retrieve cells from EMSet by cell barcodes
subset_set <- SubsetCells(em.set, cell.barcodes = random_barcodes)

}
