% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_methods.R
\name{SubsetCells}
\alias{SubsetCells}
\title{SubsetCells}
\usage{
SubsetCells(object, list)
}
\arguments{
\item{object}{An \linkS4class{AEMSet}}

\item{list}{A list of cell identifiers to subset from the \linkS4class{AEMSet}}
}
\value{
An \linkS4class{AEMSet}
}
\description{
Subset cells in the supplied list from an \linkS4class{AEMSet}.
}
