% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetmethods.R
\name{SubsetCells}
\alias{SubsetCells}
\title{SubsetCells}
\usage{
SubsetCells(object, cell_barcodes)
}
\arguments{
\item{object}{An \linkS4class{EMSet}}

\item{cell_barcodes}{A list of cell identifiers to subset from the \linkS4class{EMSet}}
}
\value{
An \linkS4class{EMSet}
}
\description{
Subset cells in the supplied list from an \linkS4class{EMSet}.
}
