% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetmethods.R
\name{SubsetCluster}
\alias{SubsetCluster}
\title{SubsetCluster}
\usage{
SubsetCluster(object, clusters)
}
\arguments{
\item{object}{A \code{\linkS4class{EMSet}}.}

\item{clusters}{Clusters to subset from the dataset.}
}
\description{
Generic for \code{\link{SubsetCluster}}.
}
