% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCEND_subsetmethods.R
\name{SubsetCluster}
\alias{SubsetCluster}
\title{SubsetCluster}
\usage{
SubsetCluster(object, clusters)
}
\arguments{
\item{object}{A \linkS4class{AEMSet} object}

\item{clusters}{Clusters to subset from the dataset.}
}
\value{
Returns an \linkS4class{AEMSet} containing only this cluster.
}
\description{
Subset a \linkS4class{AEMSet} object by cluster. Please make sure you have clustered with \code{\link{FindOptimalClusters}} before using this function.
This data is already normalised, but if you wish to recluster the data, you will need to use the RunPCA function again.
}
