% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetmethods.R
\docType{methods}
\name{SubsetCondition,EMSet-method}
\alias{SubsetCondition,EMSet-method}
\title{SubsetCondition}
\usage{
\S4method{SubsetCondition}{EMSet}(object, condition = NULL,
  subconditions = c())
}
\arguments{
\item{object}{An \code{\linkS4class{EMSet}}}

\item{condition}{Name of the condition/column you would like to subset the
\code{\linkS4class{EMSet}} by}

\item{subconditions}{List of subconditions that are stored in the condition
column, that you would like to select}
}
\value{
An \code{\linkS4class{EMSet}} containing cells that only have this condition
}
\description{
Subset cells of a specific condition(s) from an \code{\linkS4class{EMSet}}.
}
\examples{
\dontrun{
happy_emset <- SubsetCondition(em.set, condition = "emotion", 
subconditions = "happy") 
}
}
