% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetmethods.R
\name{SubsetCondition}
\alias{SubsetCondition}
\title{SubsetCondition}
\usage{
SubsetCondition(object, conditions)
}
\arguments{
\item{object}{A \linkS4class{EMSet} object}

\item{conditions}{Name of the columns which contain boolean information on whether to keep or discard a cell.}
}
\value{
An \linkS4class{EMSet} containing only this batch.
}
\description{
Subset specific conditions from a \linkS4class{EMSet} object.
}
