% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\docType{methods}
\name{SyncSlots,EMSet-method}
\alias{SyncSlots,EMSet-method}
\title{SyncSlots}
\usage{
\S4method{SyncSlots}{EMSet}(object)
}
\arguments{
\item{object}{An \linkS4class{EMSet} to synchronise}
}
\value{
Original \linkS4class{EMSet} with CellInformation, GeneInformation 
and ExpressionMatrix slots syncronised together.
}
\description{
Synchronises the data frames ExpressionMatrix, GeneInformation and 
CellInformation, to ensure everything is up to date.
}
\details{
This is important as the object will undergo a series of changes during the 
filtering and normalisation process.
}
\examples{
# Load example EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", 
"ExampleClusteredEMSet.rds"))

# Sync slots in EMSet
SyncedEMSet <- SyncSlots(EMSet)

}
