% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_utilities.R
\name{unLog2Matrix}
\alias{unLog2Matrix}
\title{unLogMatrix}
\usage{
unLog2Matrix(x)
}
\arguments{
\item{x}{Logged matrix to convert into unlogged matrix}
}
\description{
Convenience function - converts a matrix that has been converted to logcounts
to un-logged counts.
}
\examples{
# Randomly populate a matrix with 0s and 1s
normal_matrix <- matrix(sample(0:1,10*10, replace=TRUE), 10 , 10)

# Log2 + 1 values in normal_matrix
logged_matrix <- log2(normal_matrix + 1)

# Return values to normal
unlogged_matrix <- unLog2Matrix(logged_matrix)

}
