% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_utilities.R
\name{UnLog2Matrix}
\alias{UnLog2Matrix}
\title{UnLog2Matrix}
\usage{
UnLog2Matrix(expression.matrix)
}
\arguments{
\item{expression.matrix}{An expression.matrix.}
}
\value{
The expression matrix in dense matrix format.
}
\description{
Converts count values of a matrix that had undergone normalisation via 
scran's deconvolution method. This converts the log2 values + 1 pseudocount
values back into transcript count values.
}
\examples{
\dontrun{
unlogged.matrix <- Unlog2Matrix(expression.matrix)
}
}
