% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{UpdateControls}
\alias{UpdateControls}
\title{UpdateControls (Generic)}
\usage{
UpdateControls(object, controls)
}
\arguments{
\item{object}{An \linkS4class{EMSet}}

\item{controls}{Name of control group}
}
\description{
Definition for the \code{\link{UpdateControls}} function.
}
