% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{UpdateControls}
\alias{UpdateControls}
\title{UpdateControls (Generic)}
\usage{
UpdateControls(object, controls)
}
\arguments{
\item{object}{An \linkS4class{EMSet} object.}

\item{controls}{A named list containing the gene identifiers. These 
identifiers must match the identifiers used in the expression matrix. This 
list will replace any pre-existing controls.}
}
\value{
This function returns an \linkS4class{EMSet} with the updated controls.
}
\description{
Replaces the control list in a \linkS4class{EMSet} object with a new control 
list. This also recalculates the metrics associated with the 
\linkS4class{EMSet} object.
}
\details{
For best results, define your controls before you attempt any filtering. You 
can also use this function to change an EMSet into a control-less dataset.
}
\examples{
# Load EMSet
EMSet <- readRDS(system.file(package = "ascend", "extdata", "ExampleEMSet.rds"))

# List of controls to add
control_list <- list(Mt = c("MT-ATP8", "MT-ND4L", "MT-CO3"), 
Rb = c("RPL4", "RPS14", "RPS21"))

# Add controls to EMSet
UpdatedEMSet <- UpdateControls(EMSet, control_list)

}
