% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_subsetters.R
\docType{methods}
\name{useVariableGenes}
\alias{useVariableGenes}
\alias{useVariableGenes,EMSet-method}
\title{useVariableGenes}
\usage{
useVariableGenes(object, ngenes)

\S4method{useVariableGenes}{EMSet}(object, ngenes = 1500)
}
\arguments{
\item{object}{\linkS4class{EMSet}}

\item{ngenes}{Number of variable genes to use (Default: 1500 genes)}
}
\value{
EMSet reduced to specified number of variable genes.
}
\description{
Reduces an EMSet to a range of the most variable genes as determined by 
the values calculated by \code{\link[ascend:calculateCV]{calculateCV}}.
}
\examples{
# Load example EMSet
em_set <- ascend::analyzed_set

# Calculate gene variance
em_set <- calculateCV(em_set)
  
# Use 1500 of the most variable genes
em_set <- useVariableGenes(em_set, ngenes = 1500)

}
