% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascend_objects.R
\name{validateEMSet}
\alias{validateEMSet}
\title{validateEMSet}
\usage{
validateEMSet(object)
}
\arguments{
\item{object}{An EMSet that has been created, but needs validation.}
}
\description{
Validation function for the EMSet.
}
\keyword{internal}
